Created two utility files: one for generating call signs and one for cpu ship diversification. Removed generate call sign section and cpu ship diversification section from the devour scenario. This functionality now resides in external files that this scenario requires.
