+++ scripts/scenario_32_devour.lua
@@ -1,7 +1,7 @@
 -- Name: Planet Devourer
 -- Description: Stop the planet devourer before all the planets are gone
 --- 
---- Written initially for EE games on May 4th, 2023 = "May the 4th be with you"
+--- Written for EE games on May 4th, 2023 = "May the 4th be with you"
 ---
 --- Version 1
 ---
@@ -27,6 +27,8 @@
 
 require("utils.lua")
 require("place_station_scenario_utility.lua")
+require("generate_call_sign_scenario_utility.lua")
+require("cpu_ship_diversification_scenario_utility.lua")
 --	also uses supply_drop.lua
 
 --------------------
@@ -5448,2563 +5450,7 @@ function friendlyVesselDestroyed(self, instigator)
 	table.insert(friendlyVesselDestroyedValue,ship_template[tempShipType].strength)
 	--]]
 end
---		Generate call sign functions
-function generateCallSign(prefix,faction)
-	if faction == nil then
-		if prefix == nil then
-			prefix = generateCallSignPrefix()
-		end
-	else
-		if prefix == nil then
-			prefix = getFactionPrefix(faction)
-		else
-			prefix = string.format("%s %s",getFactionPrefix(faction),prefix)
-		end
-	end
-	suffix_index = suffix_index + math.random(1,3)
-	if suffix_index > 999 then 
-		suffix_index = 1
-	end
-	return string.format("%s%i",prefix,suffix_index)
-end
-function generateCallSignPrefix(length)
-	if call_sign_prefix_pool == nil then
-		call_sign_prefix_pool = {}
-		prefix_length = prefix_length + 1
-		if prefix_length > 3 then
-			prefix_length = 1
-		end
-		fillPrefixPool()
-	end
-	if length == nil then
-		length = prefix_length
-	end
-	local prefix_index = 0
-	local prefix = ""
-	for i=1,length do
-		if #call_sign_prefix_pool < 1 then
-			fillPrefixPool()
-		end
-		prefix_index = math.random(1,#call_sign_prefix_pool)
-		prefix = prefix .. call_sign_prefix_pool[prefix_index]
-		table.remove(call_sign_prefix_pool,prefix_index)
-	end
-	return prefix
-end
-function fillPrefixPool()
-	for i=1,26 do
-		table.insert(call_sign_prefix_pool,string.char(i+64))
-	end
-end
-function getFactionPrefix(faction)
-	local faction_prefix = nil
-	if faction == "Kraylor" then
-		if kraylor_names == nil then
-			setKraylorNames()
-		else
-			if #kraylor_names < 1 then
-				setKraylorNames()
-			end
-		end
-		local kraylor_name_choice = math.random(1,#kraylor_names)
-		faction_prefix = kraylor_names[kraylor_name_choice]
-		table.remove(kraylor_names,kraylor_name_choice)
-	end
-	if faction == "Exuari" then
-		if exuari_names == nil then
-			setExuariNames()
-		else
-			if #exuari_names < 1 then
-				setExuariNames()
-			end
-		end
-		local exuari_name_choice = math.random(1,#exuari_names)
-		faction_prefix = exuari_names[exuari_name_choice]
-		table.remove(exuari_names,exuari_name_choice)
-	end
-	if faction == "Ghosts" then
-		if ghosts_names == nil then
-			setGhostsNames()
-		else
-			if #ghosts_names < 1 then
-				setGhostsNames()
-			end
-		end
-		local ghosts_name_choice = math.random(1,#ghosts_names)
-		faction_prefix = ghosts_names[ghosts_name_choice]
-		table.remove(ghosts_names,ghosts_name_choice)
-	end
-	if faction == "Independent" then
-		if independent_names == nil then
-			setIndependentNames()
-		else
-			if #independent_names < 1 then
-				setIndependentNames()
-			end
-		end
-		local independent_name_choice = math.random(1,#independent_names)
-		faction_prefix = independent_names[independent_name_choice]
-		table.remove(independent_names,independent_name_choice)
-	end
-	if faction == "Human Navy" then
-		if human_names == nil then
-			setHumanNames()
-		else
-			if #human_names < 1 then
-				setHumanNames()
-			end
-		end
-		local human_name_choice = math.random(1,#human_names)
-		faction_prefix = human_names[human_name_choice]
-		table.remove(human_names,human_name_choice)
-	end
-	if faction == "Arlenians" then
-		if arlenian_names == nil then
-			setArlenianNames()
-		else
-			if #arlenian_names < 1 then
-				setArlenianNames()
-			end
-		end
-		local arlenian_name_choice = math.random(1,#arlenian_names)
-		faction_prefix = arlenian_names[arlenian_name_choice]
-		table.remove(arlenian_names,arlenian_name_choice)
-	end
-	if faction == "USN" then
-		if usn_names == nil then
-			setUsnNames()
-		else
-			if #usn_names < 1 then
-				setUsnNames()
-			end
-		end
-		local usn_name_choice = math.random(1,#usn_names)
-		faction_prefix = usn_names[usn_name_choice]
-		table.remove(usn_names,usn_name_choice)
-	end
-	if faction == "TSN" then
-		if tsn_names == nil then
-			setTsnNames()
-		else
-			if #tsn_names < 1 then
-				setTsnNames()
-			end
-		end
-		local tsn_name_choice = math.random(1,#tsn_names)
-		faction_prefix = tsn_names[tsn_name_choice]
-		table.remove(tsn_names,tsn_name_choice)
-	end
-	if faction == "CUF" then
-		if cuf_names == nil then
-			setCufNames()
-		else
-			if #cuf_names < 1 then
-				setCufNames()
-			end
-		end
-		local cuf_name_choice = math.random(1,#cuf_names)
-		faction_prefix = cuf_names[cuf_name_choice]
-		table.remove(cuf_names,cuf_name_choice)
-	end
-	if faction == "Ktlitans" then
-		if ktlitan_names == nil then
-			setKtlitanNames()
-		else
-			if #ktlitan_names < 1 then
-				setKtlitanNames()
-			end
-		end
-		local ktlitan_name_choice = math.random(1,#ktlitan_names)
-		faction_prefix = ktlitan_names[ktlitan_name_choice]
-		table.remove(ktlitan_names,ktlitan_name_choice)
-	end
-	if faction_prefix == nil then
-		faction_prefix = generateCallSignPrefix()
-	end
-	return faction_prefix
-end
-function setGhostsNames()
-	ghosts_names = {}
-	table.insert(ghosts_names,"Abstract")
-	table.insert(ghosts_names,"Ada")
-	table.insert(ghosts_names,"Assemble")
-	table.insert(ghosts_names,"Assert")
-	table.insert(ghosts_names,"Backup")
-	table.insert(ghosts_names,"BASIC")
-	table.insert(ghosts_names,"Big Iron")
-	table.insert(ghosts_names,"BigEndian")
-	table.insert(ghosts_names,"Binary")
-	table.insert(ghosts_names,"Bit")
-	table.insert(ghosts_names,"Block")
-	table.insert(ghosts_names,"Boot")
-	table.insert(ghosts_names,"Branch")
-	table.insert(ghosts_names,"BTree")
-	table.insert(ghosts_names,"Bubble")
-	table.insert(ghosts_names,"Byte")
-	table.insert(ghosts_names,"Capacitor")
-	table.insert(ghosts_names,"Case")
-	table.insert(ghosts_names,"Chad")
-	table.insert(ghosts_names,"Charge")
-	table.insert(ghosts_names,"COBOL")
-	table.insert(ghosts_names,"Collate")
-	table.insert(ghosts_names,"Compile")
-	table.insert(ghosts_names,"Control")
-	table.insert(ghosts_names,"Construct")
-	table.insert(ghosts_names,"Cycle")
-	table.insert(ghosts_names,"Data")
-	table.insert(ghosts_names,"Debug")
-	table.insert(ghosts_names,"Decimal")
-	table.insert(ghosts_names,"Decision")
-	table.insert(ghosts_names,"Default")
-	table.insert(ghosts_names,"DIMM")
-	table.insert(ghosts_names,"Displacement")
-	table.insert(ghosts_names,"Edge")
-	table.insert(ghosts_names,"Exit")
-	table.insert(ghosts_names,"Factor")
-	table.insert(ghosts_names,"Flag")
-	table.insert(ghosts_names,"Float")
-	table.insert(ghosts_names,"Flow")
-	table.insert(ghosts_names,"FORTRAN")
-	table.insert(ghosts_names,"Fullword")
-	table.insert(ghosts_names,"GIGO")
-	table.insert(ghosts_names,"Graph")
-	table.insert(ghosts_names,"Hack")
-	table.insert(ghosts_names,"Hash")
-	table.insert(ghosts_names,"Halfword")
-	table.insert(ghosts_names,"Hertz")
-	table.insert(ghosts_names,"Hexadecimal")
-	table.insert(ghosts_names,"Indicator")
-	table.insert(ghosts_names,"Initialize")
-	table.insert(ghosts_names,"Integer")
-	table.insert(ghosts_names,"Integrate")
-	table.insert(ghosts_names,"Interrupt")
-	table.insert(ghosts_names,"Java")
-	table.insert(ghosts_names,"Lisp")
-	table.insert(ghosts_names,"List")
-	table.insert(ghosts_names,"Logic")
-	table.insert(ghosts_names,"Loop")
-	table.insert(ghosts_names,"Lua")
-	table.insert(ghosts_names,"Magnetic")
-	table.insert(ghosts_names,"Mask")
-	table.insert(ghosts_names,"Memory")
-	table.insert(ghosts_names,"Mnemonic")
-	table.insert(ghosts_names,"Micro")
-	table.insert(ghosts_names,"Model")
-	table.insert(ghosts_names,"Nibble")
-	table.insert(ghosts_names,"Octal")
-	table.insert(ghosts_names,"Order")
-	table.insert(ghosts_names,"Operator")
-	table.insert(ghosts_names,"Parameter")
-	table.insert(ghosts_names,"Pascal")
-	table.insert(ghosts_names,"Pattern")
-	table.insert(ghosts_names,"Pixel")
-	table.insert(ghosts_names,"Point")
-	table.insert(ghosts_names,"Polygon")
-	table.insert(ghosts_names,"Port")
-	table.insert(ghosts_names,"Process")
-	table.insert(ghosts_names,"RAM")
-	table.insert(ghosts_names,"Raster")
-	table.insert(ghosts_names,"Rate")
-	table.insert(ghosts_names,"Redundant")
-	table.insert(ghosts_names,"Reference")
-	table.insert(ghosts_names,"Refresh")
-	table.insert(ghosts_names,"Register")
-	table.insert(ghosts_names,"Resistor")
-	table.insert(ghosts_names,"ROM")
-	table.insert(ghosts_names,"Routine")
-	table.insert(ghosts_names,"Ruby")
-	table.insert(ghosts_names,"SAAS")
-	table.insert(ghosts_names,"Sequence")
-	table.insert(ghosts_names,"Share")
-	table.insert(ghosts_names,"Silicon")
-	table.insert(ghosts_names,"SIMM")
-	table.insert(ghosts_names,"Socket")
-	table.insert(ghosts_names,"Sort")
-	table.insert(ghosts_names,"Structure")
-	table.insert(ghosts_names,"Switch")
-	table.insert(ghosts_names,"Symbol")
-	table.insert(ghosts_names,"Trace")
-	table.insert(ghosts_names,"Transistor")
-	table.insert(ghosts_names,"Value")
-	table.insert(ghosts_names,"Vector")
-	table.insert(ghosts_names,"Version")
-	table.insert(ghosts_names,"View")
-	table.insert(ghosts_names,"WYSIWYG")
-	table.insert(ghosts_names,"XOR")
-end
-function setExuariNames()
-	exuari_names = {}
-	table.insert(exuari_names,"Astonester")
-	table.insert(exuari_names,"Ametripox")
-	table.insert(exuari_names,"Bakeltevex")
-	table.insert(exuari_names,"Baropledax")
-	table.insert(exuari_names,"Batongomox")
-	table.insert(exuari_names,"Bekilvimix")
-	table.insert(exuari_names,"Benoglopok")
-	table.insert(exuari_names,"Bilontipur")
-	table.insert(exuari_names,"Bolictimik")
-	table.insert(exuari_names,"Bomagralax")
-	table.insert(exuari_names,"Buteldefex")
-	table.insert(exuari_names,"Catondinab")
-	table.insert(exuari_names,"Chatorlonox")
-	table.insert(exuari_names,"Culagromik")
-	table.insert(exuari_names,"Dakimbinix")
-	table.insert(exuari_names,"Degintalix")
-	table.insert(exuari_names,"Dimabratax")
-	table.insert(exuari_names,"Dokintifix")
-	table.insert(exuari_names,"Dotandirex")
-	table.insert(exuari_names,"Dupalgawax")
-	table.insert(exuari_names,"Ekoftupex")
-	table.insert(exuari_names,"Elidranov")
-	table.insert(exuari_names,"Fakobrovox")
-	table.insert(exuari_names,"Femoplabix")
-	table.insert(exuari_names,"Fibatralax")
-	table.insert(exuari_names,"Fomartoran")
-	table.insert(exuari_names,"Gateldepex")
-	table.insert(exuari_names,"Gamutrewal")
-	table.insert(exuari_names,"Gesanterux")
-	table.insert(exuari_names,"Gimardanax")
-	table.insert(exuari_names,"Hamintinal")
-	table.insert(exuari_names,"Holangavak")
-	table.insert(exuari_names,"Igolpafik")
-	table.insert(exuari_names,"Inoklomat")
-	table.insert(exuari_names,"Jamewtibex")
-	table.insert(exuari_names,"Jepospagox")
-	table.insert(exuari_names,"Kajortonox")
-	table.insert(exuari_names,"Kapogrinix")
-	table.insert(exuari_names,"Kelitravax")
-	table.insert(exuari_names,"Kipaldanax")
-	table.insert(exuari_names,"Kodendevex")
-	table.insert(exuari_names,"Kotelpedex")
-	table.insert(exuari_names,"Kutandolak")
-	table.insert(exuari_names,"Lakirtinix")
-	table.insert(exuari_names,"Lapoldinek")
-	table.insert(exuari_names,"Lavorbonox")
-	table.insert(exuari_names,"Letirvinix")
-	table.insert(exuari_names,"Lowibromax")
-	table.insert(exuari_names,"Makintibix")
-	table.insert(exuari_names,"Makorpohox")
-	table.insert(exuari_names,"Matoprowox")
-	table.insert(exuari_names,"Mefinketix")
-	table.insert(exuari_names,"Motandobak")
-	table.insert(exuari_names,"Nakustunux")
-	table.insert(exuari_names,"Nequivonax")
-	table.insert(exuari_names,"Nitaldavax")
-	table.insert(exuari_names,"Nobaldorex")
-	table.insert(exuari_names,"Obimpitix")
-	table.insert(exuari_names,"Owaklanat")
-	table.insert(exuari_names,"Pakendesik")
-	table.insert(exuari_names,"Pazinderix")
-	table.insert(exuari_names,"Pefoglamuk")
-	table.insert(exuari_names,"Pekirdivix")
-	table.insert(exuari_names,"Potarkadax")
-	table.insert(exuari_names,"Pulendemex")
-	table.insert(exuari_names,"Quatordunix")
-	table.insert(exuari_names,"Rakurdumux")
-	table.insert(exuari_names,"Ralombenik")
-	table.insert(exuari_names,"Regosporak")
-	table.insert(exuari_names,"Retordofox")
-	table.insert(exuari_names,"Rikondogox")
-	table.insert(exuari_names,"Rokengelex")
-	table.insert(exuari_names,"Rutarkadax")
-	table.insert(exuari_names,"Sakeldepex")
-	table.insert(exuari_names,"Setiftimix")
-	table.insert(exuari_names,"Siparkonal")
-	table.insert(exuari_names,"Sopaldanax")
-	table.insert(exuari_names,"Sudastulux")
-	table.insert(exuari_names,"Takeftebex")
-	table.insert(exuari_names,"Taliskawit")
-	table.insert(exuari_names,"Tegundolex")
-	table.insert(exuari_names,"Tekintipix")
-	table.insert(exuari_names,"Tiposhomox")
-	table.insert(exuari_names,"Tokaldapax")
-	table.insert(exuari_names,"Tomuglupux")
-	table.insert(exuari_names,"Tufeldepex")
-	table.insert(exuari_names,"Unegremek")
-	table.insert(exuari_names,"Uvendipax")
-	table.insert(exuari_names,"Vatorgopox")
-	table.insert(exuari_names,"Venitribix")
-	table.insert(exuari_names,"Vobalterix")
-	table.insert(exuari_names,"Wakintivix")
-	table.insert(exuari_names,"Wapaltunix")
-	table.insert(exuari_names,"Wekitrolax")
-	table.insert(exuari_names,"Wofarbanax")
-	table.insert(exuari_names,"Xeniplofek")
-	table.insert(exuari_names,"Yamaglevik")
-	table.insert(exuari_names,"Yakildivix")
-	table.insert(exuari_names,"Yegomparik")
-	table.insert(exuari_names,"Zapondehex")
-	table.insert(exuari_names,"Zikandelat")
-end
-function setKraylorNames()		
-	kraylor_names = {}
-	table.insert(kraylor_names,"Abroten")
-	table.insert(kraylor_names,"Ankwar")
-	table.insert(kraylor_names,"Bakrik")
-	table.insert(kraylor_names,"Belgor")
-	table.insert(kraylor_names,"Benkop")
-	table.insert(kraylor_names,"Blargvet")
-	table.insert(kraylor_names,"Bloktarg")
-	table.insert(kraylor_names,"Bortok")
-	table.insert(kraylor_names,"Bredjat")
-	table.insert(kraylor_names,"Chankret")
-	table.insert(kraylor_names,"Chatork")
-	table.insert(kraylor_names,"Chokarp")
-	table.insert(kraylor_names,"Cloprak")
-	table.insert(kraylor_names,"Coplek")
-	table.insert(kraylor_names,"Cortek")
-	table.insert(kraylor_names,"Daltok")
-	table.insert(kraylor_names,"Darpik")
-	table.insert(kraylor_names,"Dastek")
-	table.insert(kraylor_names,"Dotark")
-	table.insert(kraylor_names,"Drambok")
-	table.insert(kraylor_names,"Duntarg")
-	table.insert(kraylor_names,"Earklat")
-	table.insert(kraylor_names,"Ekmit")
-	table.insert(kraylor_names,"Fakret")
-	table.insert(kraylor_names,"Fapork")
-	table.insert(kraylor_names,"Fawtrik")
-	table.insert(kraylor_names,"Fenturp")
-	table.insert(kraylor_names,"Feplik")
-	table.insert(kraylor_names,"Figront")
-	table.insert(kraylor_names,"Floktrag")
-	table.insert(kraylor_names,"Fonkack")
-	table.insert(kraylor_names,"Fontreg")
-	table.insert(kraylor_names,"Foondrap")
-	table.insert(kraylor_names,"Frotwak")
-	table.insert(kraylor_names,"Gastonk")
-	table.insert(kraylor_names,"Gentouk")
-	table.insert(kraylor_names,"Gonpruk")
-	table.insert(kraylor_names,"Gortak")
-	table.insert(kraylor_names,"Gronkud")
-	table.insert(kraylor_names,"Hewtang")
-	table.insert(kraylor_names,"Hongtag")
-	table.insert(kraylor_names,"Hortook")
-	table.insert(kraylor_names,"Indrut")
-	table.insert(kraylor_names,"Iprant")
-	table.insert(kraylor_names,"Jakblet")
-	table.insert(kraylor_names,"Jonket")
-	table.insert(kraylor_names,"Jontot")
-	table.insert(kraylor_names,"Kandarp")
-	table.insert(kraylor_names,"Kantrok")
-	table.insert(kraylor_names,"Kiptak")
-	table.insert(kraylor_names,"Kortrant")
-	table.insert(kraylor_names,"Krontgat")
-	table.insert(kraylor_names,"Lobreck")
-	table.insert(kraylor_names,"Lokrant")
-	table.insert(kraylor_names,"Lomprok")
-	table.insert(kraylor_names,"Lutrank")
-	table.insert(kraylor_names,"Makrast")
-	table.insert(kraylor_names,"Moklahft")
-	table.insert(kraylor_names,"Morpug")
-	table.insert(kraylor_names,"Nagblat")
-	table.insert(kraylor_names,"Nokrat")
-	table.insert(kraylor_names,"Nomek")
-	table.insert(kraylor_names,"Notark")
-	table.insert(kraylor_names,"Ontrok")
-	table.insert(kraylor_names,"Orkpent")
-	table.insert(kraylor_names,"Peechak")
-	table.insert(kraylor_names,"Plogrent")
-	table.insert(kraylor_names,"Pokrint")
-	table.insert(kraylor_names,"Potarg")
-	table.insert(kraylor_names,"Prangtil")
-	table.insert(kraylor_names,"Quagbrok")
-	table.insert(kraylor_names,"Quimprill")
-	table.insert(kraylor_names,"Reekront")
-	table.insert(kraylor_names,"Ripkort")
-	table.insert(kraylor_names,"Rokust")
-	table.insert(kraylor_names,"Rontrait")
-	table.insert(kraylor_names,"Saknep")
-	table.insert(kraylor_names,"Sengot")
-	table.insert(kraylor_names,"Skitkard")
-	table.insert(kraylor_names,"Skopgrek")
-	table.insert(kraylor_names,"Sletrok")
-	table.insert(kraylor_names,"Slorknat")
-	table.insert(kraylor_names,"Spogrunk")
-	table.insert(kraylor_names,"Staklurt")
-	table.insert(kraylor_names,"Stonkbrant")
-	table.insert(kraylor_names,"Swaktrep")
-	table.insert(kraylor_names,"Tandrok")
-	table.insert(kraylor_names,"Takrost")
-	table.insert(kraylor_names,"Tonkrut")
-	table.insert(kraylor_names,"Torkrot")
-	table.insert(kraylor_names,"Trablok")
-	table.insert(kraylor_names,"Trokdin")
-	table.insert(kraylor_names,"Unkelt")
-	table.insert(kraylor_names,"Urjop")
-	table.insert(kraylor_names,"Vankront")
-	table.insert(kraylor_names,"Vintrep")
-	table.insert(kraylor_names,"Volkerd")
-	table.insert(kraylor_names,"Vortread")
-	table.insert(kraylor_names,"Wickurt")
-	table.insert(kraylor_names,"Xokbrek")
-	table.insert(kraylor_names,"Yeskret")
-	table.insert(kraylor_names,"Zacktrope")
-end
-function setIndependentNames()
-	independent_names = {}
-	table.insert(independent_names,"Akdroft")	--faux Kraylor
-	table.insert(independent_names,"Bletnik")	--faux Kraylor
-	table.insert(independent_names,"Brogfent")	--faux Kraylor
-	table.insert(independent_names,"Cruflech")	--faux Kraylor
-	table.insert(independent_names,"Dengtoct")	--faux Kraylor
-	table.insert(independent_names,"Fiklerg")	--faux Kraylor
-	table.insert(independent_names,"Groftep")	--faux Kraylor
-	table.insert(independent_names,"Hinkflort")	--faux Kraylor
-	table.insert(independent_names,"Irklesht")	--faux Kraylor
-	table.insert(independent_names,"Jotrak")	--faux Kraylor
-	table.insert(independent_names,"Kargleth")	--faux Kraylor
-	table.insert(independent_names,"Lidroft")	--faux Kraylor
-	table.insert(independent_names,"Movrect")	--faux Kraylor
-	table.insert(independent_names,"Nitrang")	--faux Kraylor
-	table.insert(independent_names,"Poklapt")	--faux Kraylor
-	table.insert(independent_names,"Raknalg")	--faux Kraylor
-	table.insert(independent_names,"Stovtuk")	--faux Kraylor
-	table.insert(independent_names,"Trongluft")	--faux Kraylor
-	table.insert(independent_names,"Vactremp")	--faux Kraylor
-	table.insert(independent_names,"Wunklesp")	--faux Kraylor
-	table.insert(independent_names,"Yentrilg")	--faux Kraylor
-	table.insert(independent_names,"Zeltrag")	--faux Kraylor
-	table.insert(independent_names,"Avoltojop")		--faux Exuari
-	table.insert(independent_names,"Bimartarax")	--faux Exuari
-	table.insert(independent_names,"Cidalkapax")	--faux Exuari
-	table.insert(independent_names,"Darongovax")	--faux Exuari
-	table.insert(independent_names,"Felistiyik")	--faux Exuari
-	table.insert(independent_names,"Gopendewex")	--faux Exuari
-	table.insert(independent_names,"Hakortodox")	--faux Exuari
-	table.insert(independent_names,"Jemistibix")	--faux Exuari
-	table.insert(independent_names,"Kilampafax")	--faux Exuari
-	table.insert(independent_names,"Lokuftumux")	--faux Exuari
-	table.insert(independent_names,"Mabildirix")	--faux Exuari
-	table.insert(independent_names,"Notervelex")	--faux Exuari
-	table.insert(independent_names,"Pekolgonex")	--faux Exuari
-	table.insert(independent_names,"Rifaltabax")	--faux Exuari
-	table.insert(independent_names,"Sobendeyex")	--faux Exuari
-	table.insert(independent_names,"Tinaftadax")	--faux Exuari
-	table.insert(independent_names,"Vadorgomax")	--faux Exuari
-	table.insert(independent_names,"Wilerpejex")	--faux Exuari
-	table.insert(independent_names,"Yukawvalak")	--faux Exuari
-	table.insert(independent_names,"Zajiltibix")	--faux Exuari
-	table.insert(independent_names,"Alter")		--faux Ghosts
-	table.insert(independent_names,"Assign")	--faux Ghosts
-	table.insert(independent_names,"Brain")		--faux Ghosts
-	table.insert(independent_names,"Break")		--faux Ghosts
-	table.insert(independent_names,"Boundary")	--faux Ghosts
-	table.insert(independent_names,"Code")		--faux Ghosts
-	table.insert(independent_names,"Compare")	--faux Ghosts
-	table.insert(independent_names,"Continue")	--faux Ghosts
-	table.insert(independent_names,"Core")		--faux Ghosts
-	table.insert(independent_names,"CRUD")		--faux Ghosts
-	table.insert(independent_names,"Decode")	--faux Ghosts
-	table.insert(independent_names,"Decrypt")	--faux Ghosts
-	table.insert(independent_names,"Device")	--faux Ghosts
-	table.insert(independent_names,"Encode")	--faux Ghosts
-	table.insert(independent_names,"Encrypt")	--faux Ghosts
-	table.insert(independent_names,"Event")		--faux Ghosts
-	table.insert(independent_names,"Fetch")		--faux Ghosts
-	table.insert(independent_names,"Frame")		--faux Ghosts
-	table.insert(independent_names,"Go")		--faux Ghosts
-	table.insert(independent_names,"IO")		--faux Ghosts
-	table.insert(independent_names,"Interface")	--faux Ghosts
-	table.insert(independent_names,"Kilo")		--faux Ghosts
-	table.insert(independent_names,"Modify")	--faux Ghosts
-	table.insert(independent_names,"Pin")		--faux Ghosts
-	table.insert(independent_names,"Program")	--faux Ghosts
-	table.insert(independent_names,"Purge")		--faux Ghosts
-	table.insert(independent_names,"Retrieve")	--faux Ghosts
-	table.insert(independent_names,"Store")		--faux Ghosts
-	table.insert(independent_names,"Unit")		--faux Ghosts
-	table.insert(independent_names,"Wire")		--faux Ghosts
-	table.insert(independent_names,"Chakak")		--faux Ktlitans
-	table.insert(independent_names,"Chakik")		--faux Ktlitans
-	table.insert(independent_names,"Chaklik")		--faux Ktlitans
-	table.insert(independent_names,"Kaklak")		--faux Ktlitans
-	table.insert(independent_names,"Kiklak")		--faux Ktlitans
-	table.insert(independent_names,"Kitpak")		--faux Ktlitans
-	table.insert(independent_names,"Kitplak")		--faux Ktlitans
-	table.insert(independent_names,"Pipklat")		--faux Ktlitans
-	table.insert(independent_names,"Piptik")		--faux Ktlitans
-end
-function setCufNames()
-	cuf_names = {}
-	table.insert(cuf_names,"Allegro")
-	table.insert(cuf_names,"Bonafide")
-	table.insert(cuf_names,"Brief Blur")
-	table.insert(cuf_names,"Byzantine Born")
-	table.insert(cuf_names,"Celeste")
-	table.insert(cuf_names,"Chosen Charter")
-	table.insert(cuf_names,"Conundrum")
-	table.insert(cuf_names,"Crazy Clef")
-	table.insert(cuf_names,"Curtail")
-	table.insert(cuf_names,"Dark Demesne")
-	table.insert(cuf_names,"Diminutive Drama")
-	table.insert(cuf_names,"Draconian Destiny")
-	table.insert(cuf_names,"Fickle Frown")
-	table.insert(cuf_names,"Final Freeze")
-	table.insert(cuf_names,"Fried Feather")
-	table.insert(cuf_names,"Frozen Flare")
-	table.insert(cuf_names,"Gaunt Gator")
-	table.insert(cuf_names,"Hidden Harpoon")
-	table.insert(cuf_names,"Intense Interest")
-	table.insert(cuf_names,"Lackadaisical")
-	table.insert(cuf_names,"Largess")
-	table.insert(cuf_names,"Ointment")
-	table.insert(cuf_names,"Plush Puzzle")
-	table.insert(cuf_names,"Slick")
-	table.insert(cuf_names,"Thumper")
-	table.insert(cuf_names,"Torpid")
-	table.insert(cuf_names,"Triple Take")
-end
-function setUsnNames()
-	usn_names = {}
-	table.insert(usn_names,"Belladonna")
-	table.insert(usn_names,"Broken Dragon")
-	table.insert(usn_names,"Burning Knave")
-	table.insert(usn_names,"Corona Flare")
-	table.insert(usn_names,"Daring the Deep")
-	table.insert(usn_names,"Dragon's Cutlass")
-	table.insert(usn_names,"Dragon's Sadness")
-	table.insert(usn_names,"Elusive Doom")
-	table.insert(usn_names,"Fast Flare")
-	table.insert(usn_names,"Flying Flare")
-	table.insert(usn_names,"Fulminate")
-	table.insert(usn_names,"Gaseous Gale")
-	table.insert(usn_names,"Golden Anger")
-	table.insert(usn_names,"Greedy Promethean")
-	table.insert(usn_names,"Happy Mynock")
-	table.insert(usn_names,"Jimi Saru")
-	table.insert(usn_names,"Jolly Roger")
-	table.insert(usn_names,"Killer's Grief")
-	table.insert(usn_names,"Mad Delight")
-	table.insert(usn_names,"Nocturnal Neptune")
-	table.insert(usn_names,"Obscure Orbiter")
-	table.insert(usn_names,"Red Rift")
-	table.insert(usn_names,"Rusty Belle")
-	table.insert(usn_names,"Silver Pearl")
-	table.insert(usn_names,"Sodden Corsair")
-	table.insert(usn_names,"Solar Sailor")
-	table.insert(usn_names,"Solar Secret")
-	table.insert(usn_names,"Sun's Grief")
-	table.insert(usn_names,"Tortuga Shadows")
-	table.insert(usn_names,"Trinity")
-	table.insert(usn_names,"Wayfaring Wind")
-end
-function setTsnNames()
-	tsn_names = {}
-	table.insert(tsn_names,"Aegis")
-	table.insert(tsn_names,"Allegiance")
-	table.insert(tsn_names,"Apollo")
-	table.insert(tsn_names,"Ares")
-	table.insert(tsn_names,"Casper")
-	table.insert(tsn_names,"Charger")
-	table.insert(tsn_names,"Dauntless")
-	table.insert(tsn_names,"Demeter")
-	table.insert(tsn_names,"Eagle")
-	table.insert(tsn_names,"Excalibur")
-	table.insert(tsn_names,"Falcon")
-	table.insert(tsn_names,"Guardian")
-	table.insert(tsn_names,"Hawk")
-	table.insert(tsn_names,"Hera")
-	table.insert(tsn_names,"Horizon")
-	table.insert(tsn_names,"Hunter")
-	table.insert(tsn_names,"Hydra")
-	table.insert(tsn_names,"Intrepid")
-	table.insert(tsn_names,"Lancer")
-	table.insert(tsn_names,"Montgomery")
-	table.insert(tsn_names,"Nemesis")
-	table.insert(tsn_names,"Osiris")
-	table.insert(tsn_names,"Pegasus")
-	table.insert(tsn_names,"Phoenix")
-	table.insert(tsn_names,"Poseidon")
-	table.insert(tsn_names,"Raven")
-	table.insert(tsn_names,"Sabre")
-	table.insert(tsn_names,"Stalker")
-	table.insert(tsn_names,"Valkyrie")
-	table.insert(tsn_names,"Viper")
-end
-function setHumanNames()
-	human_names = {}
-	table.insert(human_names,"Andromeda")
-	table.insert(human_names,"Angelica")
-	table.insert(human_names,"Artemis")
-	table.insert(human_names,"Barrier")
-	table.insert(human_names,"Beauteous")
-	table.insert(human_names,"Bliss")
-	table.insert(human_names,"Bonita")
-	table.insert(human_names,"Bounty Hunter")
-	table.insert(human_names,"Bueno")
-	table.insert(human_names,"Capitol")
-	table.insert(human_names,"Castigator")
-	table.insert(human_names,"Centurion")
-	table.insert(human_names,"Chakalaka")
-	table.insert(human_names,"Charity")
-	table.insert(human_names,"Christmas")
-	table.insert(human_names,"Chutzpah")
-	table.insert(human_names,"Constantine")
-	table.insert(human_names,"Crystal")
-	table.insert(human_names,"Dauntless")
-	table.insert(human_names,"Defiant")
-	table.insert(human_names,"Discovery")
-	table.insert(human_names,"Dorcas")
-	table.insert(human_names,"Elite")
-	table.insert(human_names,"Empathy")
-	table.insert(human_names,"Enlighten")
-	table.insert(human_names,"Enterprise")
-	table.insert(human_names,"Escape")
-	table.insert(human_names,"Exclamatory")
-	table.insert(human_names,"Faith")
-	table.insert(human_names,"Felicity")
-	table.insert(human_names,"Firefly")
-	table.insert(human_names,"Foresight")
-	table.insert(human_names,"Forthright")
-	table.insert(human_names,"Fortitude")
-	table.insert(human_names,"Frankenstein")
-	table.insert(human_names,"Gallant")
-	table.insert(human_names,"Gladiator")
-	table.insert(human_names,"Glider")
-	table.insert(human_names,"Godzilla")
-	table.insert(human_names,"Grind")
-	table.insert(human_names,"Happiness")
-	table.insert(human_names,"Hearken")
-	table.insert(human_names,"Helena")
-	table.insert(human_names,"Heracles")
-	table.insert(human_names,"Honorable Intentions")
-	table.insert(human_names,"Hope")
-	table.insert(human_names,"Inertia")
-	table.insert(human_names,"Ingenius")
-	table.insert(human_names,"Injurious")
-	table.insert(human_names,"Insight")
-	table.insert(human_names,"Insufferable")
-	table.insert(human_names,"Insurmountable")
-	table.insert(human_names,"Intractable")
-	table.insert(human_names,"Intransigent")
-	table.insert(human_names,"Jenny")
-	table.insert(human_names,"Juice")
-	table.insert(human_names,"Justice")
-	table.insert(human_names,"Jurassic")
-	table.insert(human_names,"Karma Cast")
-	table.insert(human_names,"Knockout")
-	table.insert(human_names,"Leila")
-	table.insert(human_names,"Light Fantastic")
-	table.insert(human_names,"Livid")
-	table.insert(human_names,"Lolita")
-	table.insert(human_names,"Mercury")
-	table.insert(human_names,"Moira")
-	table.insert(human_names,"Mona Lisa")
-	table.insert(human_names,"Nancy")
-	table.insert(human_names,"Olivia")
-	table.insert(human_names,"Ominous")
-	table.insert(human_names,"Oracle")
-	table.insert(human_names,"Orca")
-	table.insert(human_names,"Pandemic")
-	table.insert(human_names,"Parsimonious")
-	table.insert(human_names,"Personal Prejudice")
-	table.insert(human_names,"Porpoise")
-	table.insert(human_names,"Pristine")
-	table.insert(human_names,"Purple Passion")
-	table.insert(human_names,"Renegade")
-	table.insert(human_names,"Revelation")
-	table.insert(human_names,"Rosanna")
-	table.insert(human_names,"Rozelle")
-	table.insert(human_names,"Sainted Gramma")
-	table.insert(human_names,"Shazam")
-	table.insert(human_names,"Starbird")
-	table.insert(human_names,"Stargazer")
-	table.insert(human_names,"Stile")
-	table.insert(human_names,"Streak")
-	table.insert(human_names,"Take Flight")
-	table.insert(human_names,"Taskmaster")
-	table.insert(human_names,"The Way")
-	table.insert(human_names,"Tornado")
-	table.insert(human_names,"Trailblazer")
-	table.insert(human_names,"Trident")
-	table.insert(human_names,"Triple Threat")
-	table.insert(human_names,"Turnabout")
-	table.insert(human_names,"Undulator")
-	table.insert(human_names,"Urgent")
-	table.insert(human_names,"Victoria")
-	table.insert(human_names,"Wee Bit")
-	table.insert(human_names,"Wet Willie")
-end
-function setKtlitanNames()
-	ktlitan_names = {}
-	table.insert(ktlitan_names,"Chaklak")
-	table.insert(ktlitan_names,"Chaklit")
-	table.insert(ktlitan_names,"Chitlat")
-	table.insert(ktlitan_names,"Chitlit")
-	table.insert(ktlitan_names,"Chitpik")
-	table.insert(ktlitan_names,"Chokpit")
-	table.insert(ktlitan_names,"Choktip")
-	table.insert(ktlitan_names,"Choktot")
-	table.insert(ktlitan_names,"Chotlap")
-	table.insert(ktlitan_names,"Chotlat")
-	table.insert(ktlitan_names,"Chotlot")
-	table.insert(ktlitan_names,"Kaftlit")
-	table.insert(ktlitan_names,"Kaplak")
-	table.insert(ktlitan_names,"Kaplat")
-	table.insert(ktlitan_names,"Kichpak")
-	table.insert(ktlitan_names,"Kichpik")
-	table.insert(ktlitan_names,"Kichtak")
-	table.insert(ktlitan_names,"Kiftlat")
-	table.insert(ktlitan_names,"Kiftak")
-	table.insert(ktlitan_names,"Kiftakt")
-	table.insert(ktlitan_names,"Kiftlikt")
-	table.insert(ktlitan_names,"Kiftlit")
-	table.insert(ktlitan_names,"Kiklat")
-	table.insert(ktlitan_names,"Kiklik")
-	table.insert(ktlitan_names,"Kiklit")
-	table.insert(ktlitan_names,"Kiplit")
-	table.insert(ktlitan_names,"Kiptot")
-	table.insert(ktlitan_names,"Kitchip")
-	table.insert(ktlitan_names,"Kitchit")
-	table.insert(ktlitan_names,"Kitlaft")
-	table.insert(ktlitan_names,"Kitlak")
-	table.insert(ktlitan_names,"Kitlakt")
-	table.insert(ktlitan_names,"Kitlich")
-	table.insert(ktlitan_names,"Kitlik")
-	table.insert(ktlitan_names,"Kitpok")
-	table.insert(ktlitan_names,"Koptich")
-	table.insert(ktlitan_names,"Koptlik")
-	table.insert(ktlitan_names,"Kotplat")
-	table.insert(ktlitan_names,"Pachtik")
-	table.insert(ktlitan_names,"Paflak")
-	table.insert(ktlitan_names,"Paftak")
-	table.insert(ktlitan_names,"Paftik")
-	table.insert(ktlitan_names,"Pakchit")
-	table.insert(ktlitan_names,"Pakchok")
-	table.insert(ktlitan_names,"Paktok")
-	table.insert(ktlitan_names,"Piklit")
-	table.insert(ktlitan_names,"Piflit")
-	table.insert(ktlitan_names,"Piftik")
-	table.insert(ktlitan_names,"Pitlak")
-	table.insert(ktlitan_names,"Pochkik")
-	table.insert(ktlitan_names,"Pochkit")
-	table.insert(ktlitan_names,"Poftlit")
-	table.insert(ktlitan_names,"Pokchap")
-	table.insert(ktlitan_names,"Pokchat")
-	table.insert(ktlitan_names,"Poktat")
-	table.insert(ktlitan_names,"Poklit")
-	table.insert(ktlitan_names,"Potlak")
-	table.insert(ktlitan_names,"Tachpik")
-	table.insert(ktlitan_names,"Tachpit")
-	table.insert(ktlitan_names,"Taklit")
-	table.insert(ktlitan_names,"Talkip")
-	table.insert(ktlitan_names,"Talpik")
-	table.insert(ktlitan_names,"Taltkip")
-	table.insert(ktlitan_names,"Taltkit")
-	table.insert(ktlitan_names,"Tichpik")
-	table.insert(ktlitan_names,"Tikplit")
-	table.insert(ktlitan_names,"Tiklich")
-	table.insert(ktlitan_names,"Tiklip")
-	table.insert(ktlitan_names,"Tiklip")
-	table.insert(ktlitan_names,"Tilpit")
-	table.insert(ktlitan_names,"Tiltlit")
-	table.insert(ktlitan_names,"Tochtik")
-	table.insert(ktlitan_names,"Tochkap")
-	table.insert(ktlitan_names,"Tochpik")
-	table.insert(ktlitan_names,"Tochpit")
-	table.insert(ktlitan_names,"Tochkit")
-	table.insert(ktlitan_names,"Totlop")
-	table.insert(ktlitan_names,"Totlot")
-end
-function setArlenianNames()
-	arlenian_names = {}
-	table.insert(arlenian_names,"Balura")
-	table.insert(arlenian_names,"Baminda")
-	table.insert(arlenian_names,"Belarne")
-	table.insert(arlenian_names,"Bilanna")
-	table.insert(arlenian_names,"Calonda")
-	table.insert(arlenian_names,"Carila")
-	table.insert(arlenian_names,"Carulda")
-	table.insert(arlenian_names,"Charma")
-	table.insert(arlenian_names,"Choralle")
-	table.insert(arlenian_names,"Corlune")
-	table.insert(arlenian_names,"Damilda")
-	table.insert(arlenian_names,"Dilenda")
-	table.insert(arlenian_names,"Dorla")
-	table.insert(arlenian_names,"Elena")
-	table.insert(arlenian_names,"Emerla")
-	table.insert(arlenian_names,"Famelda")
-	table.insert(arlenian_names,"Finelle")
-	table.insert(arlenian_names,"Fontaine")
-	table.insert(arlenian_names,"Forlanne")
-	table.insert(arlenian_names,"Gendura")
-	table.insert(arlenian_names,"Gilarne")
-	table.insert(arlenian_names,"Grizelle")
-	table.insert(arlenian_names,"Hilerna")
-	table.insert(arlenian_names,"Homella")
-	table.insert(arlenian_names,"Jarille")
-	table.insert(arlenian_names,"Jindarre")
-	table.insert(arlenian_names,"Juminde")
-	table.insert(arlenian_names,"Kalena")
-	table.insert(arlenian_names,"Kimarna")
-	table.insert(arlenian_names,"Kolira")
-	table.insert(arlenian_names,"Lanerra")
-	table.insert(arlenian_names,"Lamura")
-	table.insert(arlenian_names,"Lavila")
-	table.insert(arlenian_names,"Lavorna")
-	table.insert(arlenian_names,"Lendura")
-	table.insert(arlenian_names,"Limala")
-	table.insert(arlenian_names,"Lorelle")
-	table.insert(arlenian_names,"Mavelle")
-	table.insert(arlenian_names,"Menola")
-	table.insert(arlenian_names,"Merla")
-	table.insert(arlenian_names,"Mitelle")
-	table.insert(arlenian_names,"Mivelda")
-	table.insert(arlenian_names,"Morainne")
-	table.insert(arlenian_names,"Morda")
-	table.insert(arlenian_names,"Morlena")
-	table.insert(arlenian_names,"Nadela")
-	table.insert(arlenian_names,"Naminda")
-	table.insert(arlenian_names,"Nilana")
-	table.insert(arlenian_names,"Nurelle")
-	table.insert(arlenian_names,"Panela")
-	table.insert(arlenian_names,"Pelnare")
-	table.insert(arlenian_names,"Pilera")
-	table.insert(arlenian_names,"Povelle")
-	table.insert(arlenian_names,"Quilarre")
-	table.insert(arlenian_names,"Ramila")
-	table.insert(arlenian_names,"Renatha")
-	table.insert(arlenian_names,"Rendelle")
-	table.insert(arlenian_names,"Rinalda")
-	table.insert(arlenian_names,"Riderla")
-	table.insert(arlenian_names,"Rifalle")
-	table.insert(arlenian_names,"Samila")
-	table.insert(arlenian_names,"Salura")
-	table.insert(arlenian_names,"Selinda")
-	table.insert(arlenian_names,"Simanda")
-	table.insert(arlenian_names,"Sodila")
-	table.insert(arlenian_names,"Talinda")
-	table.insert(arlenian_names,"Tamierre")
-	table.insert(arlenian_names,"Telorre")
-	table.insert(arlenian_names,"Terila")
-	table.insert(arlenian_names,"Turalla")
-	table.insert(arlenian_names,"Valerna")
-	table.insert(arlenian_names,"Vilanda")
-	table.insert(arlenian_names,"Vomera")
-	table.insert(arlenian_names,"Wanelle")
-	table.insert(arlenian_names,"Warenda")
-	table.insert(arlenian_names,"Wilena")
-	table.insert(arlenian_names,"Wodarla")
-	table.insert(arlenian_names,"Yamelda")
-	table.insert(arlenian_names,"Yelanda")
-end
 
-function stockTemplate(enemyFaction,template)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate(template)
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	return ship
-end
---------------------------------------------------------------------------------------------
---	Additional enemy ships with some modifications from the original template parameters  --
---------------------------------------------------------------------------------------------
-function farco3(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Phobos T3")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Farco 3")
-	ship:setShieldsMax(60, 40)									--stronger shields (vs 50, 40)
-	ship:setShields(60, 40)					
---				   Index,  Arc,	Dir,	Range, Cycle,	Damage
-	ship:setBeamWeapon(0,	90,	-15,	 1500,	5.0,	6.0)	--longer (vs 1200), faster (vs 8)
-	ship:setBeamWeapon(1,	90,	 15,	 1500,	5.0,	6.0)
-	local ships_key = _("scienceDB","Ships")
-	local frigate_key = _("scienceDB","Frigate")
-	local farco_key = _("scienceDB","Farco 3")
-	local phobos_key = _("scienceDB","Phobos T3")
-	local farco_3_db = queryScienceDatabase(ships_key,frigate_key,farco_key)
-	if farco_3_db == nil then
-		local frigate_db = queryScienceDatabase(ships_key,frigate_key)
-		if frigate_db ~= nil then	--added for translation issues
-			frigate_db:addEntry(farco_key)
-			farco_3_db = queryScienceDatabase(ships_key,frigate_key,farco_key)
-			local tube_key = _("scienceDB","Tube -1")
-			local tube2_key = _("scienceDB","Tube 1")
-			local load_val = _("scienceDB","60 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,frigate_key,phobos_key),	--base ship database entry
-				farco_3_db,		--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Farco models are evolutionary changes to the Phobos T3. In the case of the Farco 3, the beams are longer and faster and the shields are slightly stronger."),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},		--torpedo tube direction and load speed
-				},
-				nil		--jump range
-			)
-		end
-	end
-	return ship
-end
-function farco5(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Phobos T3")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Farco 5")
-	ship:setShieldsMax(60, 40)				--stronger shields (vs 50, 40)
-	ship:setShields(60, 40)	
-	ship:setTubeLoadTime(0,30)				--faster (vs 60)
-	ship:setTubeLoadTime(0,30)				
-	local ships_key = _("scienceDB","Ships")
-	local frigate_key = _("scienceDB","Frigate")
-	local farco_key = _("scienceDB","Farco 5")
-	local phobos_key = _("scienceDB","Phobos T3")
-	local farco_5_db = queryScienceDatabase(ships_key,frigate_key,farco_key)
-	if farco_5_db == nil then
-		local frigate_db = queryScienceDatabase(ships_key,frigate_key)
-		if frigate_db ~= nil then	--added for translation issues
-			frigate_db:addEntry(farco_key)
-			farco_5_db = queryScienceDatabase(ships_key,frigate_key,farco_key)
-			local tube_key = _("scienceDB","Tube -1")
-			local tube2_key = _("scienceDB","Tube 1")
-			local load_val = _("scienceDB","30 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,frigate_key,phobos_key),	--base ship database entry
-				farco_5_db,		--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Farco models are evolutionary changes to the Phobos T3. In the case of the Farco 5, the tubes load faster and the shields are slightly stronger."),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},		--torpedo tube direction and load speed
-				},
-				nil		--jump range
-			)
-		end
-	end
-	return ship
-end
-function farco8(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Phobos T3")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Farco 8")
-	ship:setShieldsMax(80, 50)				--stronger shields (vs 50, 40)
-	ship:setShields(80, 50)	
---				   Index,  Arc,	Dir,	Range, Cycle,	Damage
-	ship:setBeamWeapon(0,	90,	-15,	 1500,	5.0,	6.0)	--longer (vs 1200), faster (vs 8)
-	ship:setBeamWeapon(1,	90,	 15,	 1500,	5.0,	6.0)
-	ship:setTubeLoadTime(0,30)				--faster (vs 60)
-	ship:setTubeLoadTime(0,30)				
-	local ships_key = _("scienceDB","Ships")
-	local frigate_key = _("scienceDB","Frigate")
-	local farco_key = _("scienceDB","Farco 8")
-	local phobos_key = _("scienceDB","Phobos T3")
-	local farco_8_db = queryScienceDatabase(ships_key,frigate_key,farco_key)
-	if farco_8_db == nil then
-		local frigate_db = queryScienceDatabase(ships_key,frigate_key)
-		if frigate_db ~= nil then	--added for translation issues
-			frigate_db:addEntry(farco_key)
-			farco_8_db = queryScienceDatabase(ships_key,frigate_key,farco_key)
-			local tube_key = _("scienceDB","Tube -1")
-			local tube2_key = _("scienceDB","Tube 1")
-			local load_val = _("scienceDB","30 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,frigate_key,phobos_key),	--base ship database entry
-				farco_8_db,		--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Farco models are evolutionary changes to the Phobos T3. In the case of the Farco 8, the beams are longer and faster, the tubes load faster and the shields are stronger."),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},		--torpedo tube direction and load speed
-				},
-				nil		--jump range
-			)
-		end
-	end
-	return ship
-end
-function farco11(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Phobos T3")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Farco 11")
-	ship:setShieldsMax(80, 50)				--stronger shields (vs 50, 40)
-	ship:setShields(80, 50)	
-	ship:setRotationMaxSpeed(15)								--faster maneuver (vs 10)
---				   Index,  Arc,	Dir,	Range, Cycle,	Damage
-	ship:setBeamWeapon(0,	90,	-15,	 1500,	5.0,	6.0)	--longer (vs 1200), faster (vs 8)
-	ship:setBeamWeapon(1,	90,	 15,	 1500,	5.0,	6.0)
-	ship:setBeamWeapon(2,	20,	  0,	 1800,	5.0,	4.0)	--additional sniping beam
-	local ships_key = _("scienceDB","Ships")
-	local frigate_key = _("scienceDB","Frigate")
-	local farco_key = _("scienceDB","Farco 11")
-	local phobos_key = _("scienceDB","Phobos T3")
-	local farco_11_db = queryScienceDatabase(ships_key,frigate_key,farco_key)
-	if farco_11_db == nil then
-		local frigate_db = queryScienceDatabase(ships_key,frigate_key)
-		if frigate_db ~= nil then	--added for translation issues
-			frigate_db:addEntry(farco_key)
-			farco_11_db = queryScienceDatabase(ships_key,frigate_key,farco_key)
-			local tube_key = _("scienceDB","Tube -1")
-			local tube2_key = _("scienceDB","Tube 1")
-			local load_val = _("scienceDB","60 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,frigate_key,phobos_key),	--base ship database entry
-				farco_11_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Farco models are evolutionary changes to the Phobos T3. In the case of the Farco 11, the maneuver speed is faster, the beams are longer and faster, there's an added longer sniping beam and the shields are stronger."),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},		--torpedo tube direction and load speed
-				},
-				nil		--jump range
-			)
-		end
-	end
-	return ship
-end
-function farco13(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Phobos T3")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Farco 13")
-	ship:setShieldsMax(90, 70)				--stronger shields (vs 50, 40)
-	ship:setShields(90, 70)	
-	ship:setRotationMaxSpeed(15)								--faster maneuver (vs 10)
---				   Index,  Arc,	Dir,	Range, Cycle,	Damage
-	ship:setBeamWeapon(0,	90,	-15,	 1500,	5.0,	6.0)	--longer (vs 1200), faster (vs 8)
-	ship:setBeamWeapon(1,	90,	 15,	 1500,	5.0,	6.0)
-	ship:setBeamWeapon(2,	20,	  0,	 1800,	5.0,	4.0)	--additional sniping beam
-	ship:setTubeLoadTime(0,30)				--faster (vs 60)
-	ship:setTubeLoadTime(0,30)				
-	ship:setWeaponStorageMax("Homing",16)						--more (vs 6)
-	ship:setWeaponStorage("Homing", 16)		
-	ship:setWeaponStorageMax("HVLI",30)							--more (vs 20)
-	ship:setWeaponStorage("HVLI", 30)
-	local ships_key = _("scienceDB","Ships")
-	local frigate_key = _("scienceDB","Frigate")
-	local farco_key = _("scienceDB","Farco 13")
-	local phobos_key = _("scienceDB","Phobos T3")
-	local farco_13_db = queryScienceDatabase(ships_key,frigate_key,farco_key)
-	if farco_13_db == nil then
-		local frigate_db = queryScienceDatabase(ships_key,frigate_key)
-		if frigate_db ~= nil then	--added for translation issues
-			frigate_db:addEntry(farco_key)
-			farco_13_db = queryScienceDatabase(ships_key,frigate_key,farco_key)
-			local tube_key = _("scienceDB","Tube -1")
-			local tube2_key = _("scienceDB","Tube 1")
-			local load_val = _("scienceDB","30 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,frigate_key,phobos_key),	--base ship database entry
-				farco_13_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Farco models are evolutionary changes to the Phobos T3. In the case of the Farco 13, the maneuver speed is faster, the beams are longer and faster, there's an added longer sniping beam, the tubes load faster, there are more missiles and the shields are stronger."),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},		--torpedo tube direction and load speed
-				},
-				nil		--jump range
-			)
-		end
-	end
-	return ship
-end
-function whirlwind(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Storm")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Whirlwind")
-	ship:setWeaponTubeCount(9)					--more (vs 5)
-	ship:setWeaponTubeDirection(0,-90)			--3 left, 3 right, 3 front (vs 5 front)	
-	ship:setWeaponTubeDirection(1,-92)				
-	ship:setWeaponTubeDirection(2,-88)				
-	ship:setWeaponTubeDirection(3, 90)				
-	ship:setWeaponTubeDirection(4, 92)				
-	ship:setWeaponTubeDirection(5, 88)				
-	ship:setWeaponTubeDirection(6,  0)				
-	ship:setWeaponTubeDirection(7,  2)				
-	ship:setWeaponTubeDirection(8, -2)				
-	ship:setWeaponStorageMax("Homing",36)						--more (vs 15)
-	ship:setWeaponStorage("Homing", 36)		
-	ship:setWeaponStorageMax("HVLI",36)							--more (vs 15)
-	ship:setWeaponStorage("HVLI", 36)
-	local ships_key = _("scienceDB","Ships")
-	local frigate_key = _("scienceDB","Frigate")
-	local whirlwind_key = _("scienceDB","Whirlwind")
-	local storm_key = _("scienceDB","Storm")
-	local whirlwind_db = queryScienceDatabase(ships_key,frigate_key,whirlwind_key)
-	if whirlwind_db == nil then
-		local frigate_db = queryScienceDatabase(ships_key,frigate_key)
-		if frigate_db ~= nil then	--added for translation issues
-			frigate_db:addEntry(whirlwind_key)
-			whirlwind_db = queryScienceDatabase(ships_key,frigate_key,whirlwind_key)
-			local tube_key = _("scienceDB","Tube -90")
-			local tube2_key = _("scienceDB","Tube -92")
-			local tube3_key = _("scienceDB","Tube -88")
-			local tube4_key = _("scienceDB","Tube  90")
-			local tube5_key = _("scienceDB","Tube  92")
-			local tube6_key = _("scienceDB","Tube  88")
-			local tube7_key = _("scienceDB","Tube   0")
-			local tube8_key = _("scienceDB","Tube   2")
-			local tube9_key = _("scienceDB","Tube  -2")
-			local load_val = _("scienceDB","15 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,frigate_key,storm_key),	--base ship database entry
-				whirlwind_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Whirlwind, another heavy artillery cruiser, takes the Storm and adds tubes and missiles. It's as if the Storm swallowed a Pirahna and grew gills. Expect to see missiles, lots of missiles"),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube3_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube4_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube5_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube6_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube7_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube8_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube9_key, value = load_val},	--torpedo tube direction and load speed
-				},
-				nil		--jump range
-			)
-		end
-	end
-	return ship
-end
-function phobosR2(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Phobos T3")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Phobos R2")
-	ship:setWeaponTubeCount(1)			--one tube (vs 2)
-	ship:setWeaponTubeDirection(0,0)	
-	ship:setImpulseMaxSpeed(55)			--slower impulse (vs 60)
-	ship:setRotationMaxSpeed(15)		--faster maneuver (vs 10)
-	local ships_key = _("scienceDB","Ships")
-	local frigate_key = _("scienceDB","Frigate")
-	local phobos_r2_key = _("scienceDB","Phobos R2")
-	local phobos_key = _("scienceDB","Phobos T3")
-	local phobos_r2_db = queryScienceDatabase(ships_key,frigate_key,phobos_r2_key)
-	if phobos_r2_db == nil then
-		local frigate_db = queryScienceDatabase(ships_key,frigate_key)
-		if frigate_db ~= nil then	--added for translation issues
-			frigate_db:addEntry(phobos_r2_key)
-			phobos_r2_db = queryScienceDatabase(ships_key,frigate_key,phobos_r2_key)
-			local tube_key = _("scienceDB","Tube 0")
-			local load_val = _("scienceDB","60 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,frigate_key,phobos_key),	--base ship database entry
-				phobos_r2_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Phobos R2 model is very similar to the Phobos T3. It's got a faster turn speed, but only one missile tube"),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-				},
-				nil
-			)
-		end
-	end
-	return ship
-end
-function hornetMV52(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("MT52 Hornet")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("MV52 Hornet")
-	ship:setBeamWeapon(0, 30, 0, 1000.0, 4.0, 3.0)	--longer and stronger beam (vs 700 & 2)
-	ship:setRotationMaxSpeed(31)					--faster maneuver (vs 30)
-	ship:setImpulseMaxSpeed(130)					--faster impulse (vs 120)
-	local ships_key = _("scienceDB","Ships")
-	local starfighter_key = _("scienceDB","Starfighter")
-	local mv52_hornet_key = _("scienceDB","MV52 Hornet")
-	local hornet_key = _("scienceDB","MT52 Hornet")
-	local hornet_mv52_db = queryScienceDatabase(ships_key,starfighter_key,mv52_hornet_key)
-	if hornet_mv52_db == nil then
-		local starfighter_db = queryScienceDatabase(ships_key,starfighter_key)
-		if starfighter_db ~= nil then	--added for translation issues
-			starfighter_db:addEntry(mv52_hornet_key)
-			hornet_mv52_db = queryScienceDatabase(ships_key,starfighter_key,mv52_hornet_key)
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,starfighter_key,hornet_key),	--base ship database entry
-				hornet_mv52_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The MV52 Hornet is very similar to the MT52 and MU52 models. The beam does more damage than both of the other Hornet models, it's max impulse speed is faster than both of the other Hornet models, it turns faster than the MT52, but slower than the MU52"),
-				nil,
-				nil
-			)
-		end
-	end
-	return ship
-end
-function k2fighter(enemyFaction)
-	local k2_key = _("scienceDB","K2 Fighter")
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Ktlitan Fighter")
-	ship:setTypeName(k2_key)
-	ship:setBeamWeapon(0, 60, 0, 1200.0, 2.5, 6)	--beams cycle faster (vs 4.0)
-	ship:setHullMax(65)								--weaker hull (vs 70)
-	ship:setHull(65)
-	local ships_key = _("scienceDB","Ships")
-	local no_class_key = _("scienceDB","No Class")
-	local ktlitan_key = _("scienceDB","Ktlitan Fighter")
-	local k2_fighter_db = queryScienceDatabase(ships_key,no_class_key,k2_key)
-	if k2_fighter_db == nil then
-		local no_class_db = queryScienceDatabase(ships_key,no_class_key)
-		if no_class_db ~= nil then	--added for translation issues
-			no_class_db:addEntry(k2_key)
-			k2_fighter_db = queryScienceDatabase(ships_key,no_class_key,k2_key)
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,no_class_key,ktlitan_key),	--base ship database entry
-				k2_fighter_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","Enterprising designers published this design specification based on salvaged Ktlitan Fighters. Comparatively, it's got beams that cycle faster, but the hull is a bit weaker."),
-				nil,
-				nil		--jump range
-			)
-		end
-	end
-	return ship
-end	
-function k3fighter(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Ktlitan Fighter")
-	ship:setTypeName("K3 Fighter")
-	ship:setBeamWeapon(0, 60, 0, 1200.0, 2.5, 9)	--beams cycle faster and damage more (vs 4.0 & 6)
-	ship:setHullMax(60)								--weaker hull (vs 70)
-	ship:setHull(60)
-	local ships_key = _("scienceDB","Ships")
-	local no_class_key = _("scienceDB","No Class")
-	local k3_key = _("scienceDB","K3 Fighter")
-	local ktlitan_key = _("scienceDB","Ktlitan Fighter")
-	local k3_fighter_db = queryScienceDatabase(ships_key,no_class_key,k3_key)
-	if k3_fighter_db == nil then
-		local no_class_db = queryScienceDatabase(ships_key,no_class_key)
-		if no_class_db ~= nil then	--added for translation issues
-			no_class_db:addEntry(k3_key)
-			k3_fighter_db = queryScienceDatabase(ships_key,no_class_key,k3_key)
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,no_class_key,ktlitan_key),	--base ship database entry
-				k3_fighter_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","Enterprising designers published this design specification based on salvaged Ktlitan Fighters. Comparatively, it's got beams that are stronger and that cycle faster, but the hull is weaker."),
-				nil,
-				nil		--jump range
-			)
-		end
-	end
-	return ship
-end	
-function waddle5(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Adder MK5")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Waddle 5")
-	ship:setWarpDrive(true)
---				   Index,  Arc,	  Dir, Range, Cycle,	Damage
-	ship:setBeamWeapon(2,	70,	  -30,	 600,	5.0,	2.0)	--adjust beam direction to match starboard side (vs -35)
-	local ships_key = _("scienceDB","Ships")
-	local starfighter_key = _("scienceDB","Starfighter")
-	local waddle5_key = _("scienceDB","Waddle 5")
-	local adder_key = _("scienceDB","Adder MK5")
-	local waddle_5_db = queryScienceDatabase(ships_key,starfighter_key,waddle5_key)
-	if waddle_5_db == nil then
-		local starfighter_db = queryScienceDatabase(ships_key,starfighter_key)
-		if starfighter_db ~= nil then	--added for translation issues
-			starfighter_db:addEntry(waddle5_key)
-			waddle_5_db = queryScienceDatabase(ships_key,starfighter_key,waddle5_key)
-			local tube_key = _("scienceDB","Small tube 0")
-			local load_val = _("scienceDB","15 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,starfighter_key,adder_key),	--base ship database entry
-				waddle_5_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","Conversions R Us purchased a number of Adder MK 5 ships at auction and added warp drives to them to produce the Waddle 5"),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-				},
-				nil		--jump range
-			)
-		end
-	end
-	return ship
-end
-function jade5(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Adder MK5")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Jade 5")
-	ship:setJumpDrive(true)
-	ship:setJumpDriveRange(5000,35000)
---				   Index,  Arc,	  Dir, Range, Cycle,	Damage
-	ship:setBeamWeapon(2,	70,	  -30,	 600,	5.0,	2.0)	--adjust beam direction to match starboard side (vs -35)
-	local ships_key = _("scienceDB","Ships")
-	local starfighter_key = _("scienceDB","Starfighter")
-	local jade5_key = _("scienceDB","Jade 5")
-	local adder_key = _("scienceDB","Adder MK5")
-	local jade_5_db = queryScienceDatabase(ships_key,starfighter_key,jade5_key)
-	if jade_5_db == nil then
-		local starfighter_db = queryScienceDatabase(ships_key,starfighter_key)
-		if starfighter_db ~= nil then	--added for translation issues
-			starfighter_db:addEntry(jade5_key)
-			jade_5_db = queryScienceDatabase(ships_key,starfighter_key,jade5_key)
-			local tube_key = _("scienceDB","Small tube 0")
-			local load_val = _("scienceDB","15 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,starfighter_key,adder_key),	--base ship database entry
-				jade_5_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","Conversions R Us purchased a number of Adder MK 5 ships at auction and added jump drives to them to produce the Jade 5"),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-				},
-				"5 - 35 U"		--jump range
-			)
-		end
-	end
-	return ship
-end
-function droneLite(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Ktlitan Drone")
-	ship:setTypeName("Lite Drone")
-	ship:setHullMax(20)					--weaker hull (vs 30)
-	ship:setHull(20)
-	ship:setImpulseMaxSpeed(130)		--faster impulse (vs 120)
-	ship:setRotationMaxSpeed(20)		--faster maneuver (vs 10)
-	ship:setBeamWeapon(0,40,0,600,4,4)	--weaker (vs 6) beam
-	local ships_key = _("scienceDB","Ships")
-	local no_class_key = _("scienceDB","No Class")
-	local lite_drone_key = _("scienceDB","Lite Drone")
-	local ktlitan_key = _("scienceDB","Ktlitan Drone")
-	local drone_lite_db = queryScienceDatabase(ships_key,no_class_key,lite_drone_key)
-	if drone_lite_db == nil then
-		local no_class_db = queryScienceDatabase(ships_key,no_class_key)
-		if no_class_db ~= nil then	--added for translation issues
-			no_class_db:addEntry(lite_drone_key)
-			drone_lite_db = queryScienceDatabase(ships_key,no_class_key,lite_drone_key)
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,no_class_key,ktlitan_key),	--base ship database entry
-				drone_lite_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The light drone was pieced together from scavenged parts of various damaged Ktlitan drones. Compared to the Ktlitan drone, the lite drone has a weaker hull, and a weaker beam, but a faster turn and impulse speed"),
-				nil,
-				nil
-			)
-		end
-	end
-	return ship
-end
-function droneHeavy(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Ktlitan Drone")
-	ship:setTypeName("Heavy Drone")
-	ship:setHullMax(40)					--stronger hull (vs 30)
-	ship:setHull(40)
-	ship:setImpulseMaxSpeed(110)		--slower impulse (vs 120)
-	ship:setBeamWeapon(0,40,0,600,4,8)	--stronger (vs 6) beam
-	local ships_key = _("scienceDB","Ships")
-	local no_class_key = _("scienceDB","No Class")
-	local heavy_drone_key = _("scienceDB","Heavy Drone")
-	local ktlitan_key = _("scienceDB","Ktlitan Drone")
-	local drone_heavy_db = queryScienceDatabase(ships_key,no_class_key,heavy_drone_key)
-	if drone_heavy_db == nil then
-		local no_class_db = queryScienceDatabase(ships_key,no_class_key)
-		if no_class_db ~= nil then	--added for translation issues
-			no_class_db:addEntry(heavy_drone_key)
-			drone_heavy_db = queryScienceDatabase(ships_key,no_class_key,heavy_drone_key)
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,no_class_key,ktlitan_key),	--base ship database entry
-				drone_heavy_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The heavy drone has a stronger hull and a stronger beam than the normal Ktlitan Drone, but it also moves slower"),
-				nil,
-				nil
-			)
-		end
-	end
-	return ship
-end
-function droneJacket(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Ktlitan Drone")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Jacket Drone")
-	ship:setShieldsMax(20)				--stronger shields (vs none)
-	ship:setShields(20)
-	ship:setImpulseMaxSpeed(110)		--slower impulse (vs 120)
-	ship:setBeamWeapon(0,40,0,600,4,4)	--weaker (vs 6) beam
-	local ships_key = _("scienceDB","Ships")
-	local no_class_key = _("scienceDB","No Class")
-	local jacket_drone_key = _("scienceDB","Jacket Drone")
-	local ktlitan_key = _("scienceDB","Ktlitan Drone")
-	local drone_jacket_db = queryScienceDatabase(ships_key,no_class_key,jacket_drone_key)
-	if drone_jacket_db == nil then
-		local no_class_db = queryScienceDatabase(ships_key,no_class_key)
-		if no_class_db ~= nil then	--added for translation issues
-			no_class_db:addEntry(jacket_drone_key)
-			drone_jacket_db = queryScienceDatabase(ships_key,no_class_key,jacket_drone_key)
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,no_class_key,ktlitan_key),	--base ship database entry
-				drone_jacket_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Jacket Drone is a Ktlitan Drone with a shield. It's also slightly slower and has a slightly weaker beam due to the energy requirements of the added shield"),
-				nil,
-				nil
-			)
-		end
-	end
-	return ship
-end
-function wzLindworm(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("WX-Lindworm")
-	ship:setTypeName("WZ-Lindworm")
-	ship:setWeaponStorageMax("Nuke",2)		--more nukes (vs 0)
-	ship:setWeaponStorage("Nuke",2)
-	ship:setWeaponStorageMax("Homing",4)	--more homing (vs 1)
-	ship:setWeaponStorage("Homing",4)
-	ship:setWeaponStorageMax("HVLI",12)		--more HVLI (vs 6)
-	ship:setWeaponStorage("HVLI",12)
-	ship:setRotationMaxSpeed(12)			--slower maneuver (vs 15)
-	ship:setHullMax(45)						--weaker hull (vs 50)
-	ship:setHull(45)
-	local ships_key = _("scienceDB","Ships")
-	local starfighter_key = _("scienceDB","Starfighter")
-	local wzlindworm_key = _("scienceDB","WZ-Lindworm")
-	local worm_key = _("scienceDB","WX-Lindworm")
-	local wz_lindworm_db = queryScienceDatabase(ships_key,starfighter_key,wzlindworm_key)
-	if wz_lindworm_db == nil then
-		local starfighter_db = queryScienceDatabase(ships_key,starfighter_key)
-		if starfighter_db ~= nil then	--added for translation issues
-			starfighter_db:addEntry(wzlindworm_key)
-			wz_lindworm_db = queryScienceDatabase(ships_key,starfighter_key,wzlindworm_key)
-			local tube_key = _("scienceDB","Small tube 0")
-			local tube2_key = _("scienceDB","Small tube 1")
-			local tube3_key = _("scienceDB","Small tube -1")
-			local load_val = _("scienceDB","15 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,starfighter_key,worm_key),	--base ship database entry
-				wz_lindworm_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The WZ-Lindworm is essentially the stock WX-Lindworm with more HVLIs, more homing missiles and added nukes. They had to remove some of the armor to get the additional missiles to fit, so the hull is weaker. Also, the WZ turns a little more slowly than the WX. This little bomber packs quite a whallop."),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube3_key, value = load_val},	--torpedo tube direction and load speed
-				},
-				nil
-			)
-		end
-	end
-	return ship
-end
-function tempest(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Piranha F12")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Tempest")
-	ship:setWeaponTubeCount(10)						--four more tubes (vs 6)
-	ship:setWeaponTubeDirection(0, -88)				--5 per side
-	ship:setWeaponTubeDirection(1, -89)				--slight angle spread
-	ship:setWeaponTubeDirection(3,  88)				--3 for HVLI each side
-	ship:setWeaponTubeDirection(4,  89)				--2 for homing and nuke each side
-	ship:setWeaponTubeDirection(6, -91)				
-	ship:setWeaponTubeDirection(7, -92)				
-	ship:setWeaponTubeDirection(8,  91)				
-	ship:setWeaponTubeDirection(9,  92)				
-	ship:setWeaponTubeExclusiveFor(7,"HVLI")
-	ship:setWeaponTubeExclusiveFor(9,"HVLI")
-	ship:setWeaponStorageMax("Homing",16)			--more (vs 6)
-	ship:setWeaponStorage("Homing", 16)				
-	ship:setWeaponStorageMax("Nuke",8)				--more (vs 0)
-	ship:setWeaponStorage("Nuke", 8)				
-	ship:setWeaponStorageMax("HVLI",34)				--more (vs 20)
-	ship:setWeaponStorage("HVLI", 34)
-	local ships_key = _("scienceDB","Ships")
-	local frigate_key = _("scienceDB","Frigate")
-	local tempest_key = _("scienceDB","Tempest")
-	local pirahna_key = _("scienceDB","Piranha F12")
-	local tempest_db = queryScienceDatabase(ships_key,frigate_key,tempest_key)
-	if tempest_db == nil then
-		local frigate_db = queryScienceDatabase(ships_key,frigate_key)
-		if frigate_db ~= nil then	--added for translation issues
-			frigate_db:addEntry(tempest_key)
-			tempest_db = queryScienceDatabase(ships_key,frigate_key,tempest_key)
-			local tube_key = _("scienceDB","Large tube -88")
-			local tube2_key = _("scienceDB","Tube -89")
-			local tube3_key = _("scienceDB","Large tube -90")
-			local tube4_key = _("scienceDB","Large tube 88")
-			local tube5_key = _("scienceDB","Tube 89")
-			local tube6_key = _("scienceDB","Large tube 90")
-			local tube7_key = _("scienceDB","Tube -91")
-			local tube8_key = _("scienceDB","Tube -92")
-			local tube9_key = _("scienceDB","Tube 91")
-			local tube10_key = _("scienceDB","Tube 92")
-			local load_val = _("scienceDB","15 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,frigate_key,pirahna_key),	--base ship database entry
-				tempest_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","Loosely based on the Piranha F12 model, the Tempest adds four more broadside tubes (two on each side), more HVLIs, more Homing missiles and 8 Nukes. The Tempest can strike fear into the hearts of your enemies. Get yourself one today!"),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},		--torpedo tube direction and load speed
-					{key = tube3_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube4_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube5_key, value = load_val},		--torpedo tube direction and load speed
-					{key = tube6_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube7_key, value = load_val},		--torpedo tube direction and load speed
-					{key = tube8_key, value = load_val},		--torpedo tube direction and load speed
-					{key = tube9_key, value = load_val},		--torpedo tube direction and load speed
-					{key = tube10_key, value = load_val},		--torpedo tube direction and load speed
-				},
-				nil
-			)
-		end
-	end
-	return ship
-end
-function enforcer(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Blockade Runner")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Enforcer")
-	ship:setRadarTrace("ktlitan_destroyer.png")			--different radar trace
-	ship:setWarpDrive(true)										--warp (vs none)
-	ship:setWarpSpeed(600)
-	ship:setImpulseMaxSpeed(100)								--faster impulse (vs 60)
-	ship:setRotationMaxSpeed(20)								--faster maneuver (vs 15)
-	ship:setShieldsMax(200,100,100)								--stronger shields (vs 100,150)
-	ship:setShields(200,100,100)					
-	ship:setHullMax(100)										--stronger hull (vs 70)
-	ship:setHull(100)
---				   Index,  Arc,	  Dir, Range,	Cycle,	Damage
-	ship:setBeamWeapon(0,	30,	    5,	1500,		6,		10)	--narrower (vs 60), longer (vs 1000), stronger (vs 8)
-	ship:setBeamWeapon(1,	30,	   -5,	1500,		6,		10)
-	ship:setBeamWeapon(2,	 0,	    0,	   0,		0,		 0)	--fewer (vs 4)
-	ship:setBeamWeapon(3,	 0,	    0,	   0,		0,		 0)
-	ship:setWeaponTubeCount(3)									--more (vs 0)
-	ship:setTubeSize(0,"large")									--large (vs normal)
-	ship:setWeaponTubeDirection(1,-15)				
-	ship:setWeaponTubeDirection(2, 15)				
-	ship:setTubeLoadTime(0,18)
-	ship:setTubeLoadTime(1,12)
-	ship:setTubeLoadTime(2,12)			
-	ship:setWeaponStorageMax("Homing",18)						--more (vs 0)
-	ship:setWeaponStorage("Homing", 18)
-	local ships_key = _("scienceDB","Ships")
-	local frigate_key = _("scienceDB","Frigate")
-	local enforcer_key = _("scienceDB","Enforcer")
-	local blockade_runner_key = _("scienceDB","Blockade Runner")
-	local enforcer_db = queryScienceDatabase(ships_key,frigate_key,enforcer_key)
-	if enforcer_db == nil then
-		local frigate_db = queryScienceDatabase(ships_key,frigate_key)
-		if frigate_db ~= nil then	--added for translation issues
-			frigate_db:addEntry(enforcer_key)
-			enforcer_db = queryScienceDatabase(ships_key,frigate_key,enforcer_key)
-			local tube_key = _("scienceDB","Large tube 0")
-			local tube2_key = _("scienceDB","Tube -15")
-			local tube3_key = _("scienceDB","Tube 15")
-			local load_val = _("scienceDB","18 sec")
-			local load2_val = _("scienceDB","12 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,frigate_key,blockade_runner_key),	--base ship database entry
-				enforcer_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Enforcer is a highly modified Blockade Runner. A warp drive was added and impulse engines boosted along with turning speed. Three missile tubes were added to shoot homing missiles, large ones straight ahead. Stronger shields and hull. Removed rear facing beams and strengthened front beams."),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load2_val},		--torpedo tube direction and load speed
-					{key = tube3_key, value = load2_val},		--torpedo tube direction and load speed
-				},
-				nil
-			)
-			enforcer_db:setImage("radar/ktlitan_destroyer.png")		--override default radar image
-		end
-	end
-	return ship		
-end
-function predator(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Piranha F8")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Predator")
-	ship:setShieldsMax(100,100)									--stronger shields (vs 30,30)
-	ship:setShields(100,100)					
-	ship:setHullMax(80)											--stronger hull (vs 70)
-	ship:setHull(80)
-	ship:setImpulseMaxSpeed(65)									--faster impulse (vs 40)
-	ship:setRotationMaxSpeed(15)								--faster maneuver (vs 6)
-	ship:setJumpDrive(true)
-	ship:setJumpDriveRange(5000,35000)			
---				   Index,  Arc,	  Dir, Range,	Cycle,	Damage
-	ship:setBeamWeapon(0,	90,	    0,	1000,		6,		 4)	--more (vs 0)
-	ship:setBeamWeapon(1,	90,	  180,	1000,		6,		 4)	
-	ship:setWeaponTubeCount(8)									--more (vs 3)
-	ship:setWeaponTubeDirection(0,-60)				
-	ship:setWeaponTubeDirection(1,-90)				
-	ship:setWeaponTubeDirection(2,-90)				
-	ship:setWeaponTubeDirection(3, 60)				
-	ship:setWeaponTubeDirection(4, 90)				
-	ship:setWeaponTubeDirection(5, 90)				
-	ship:setWeaponTubeDirection(6,-120)				
-	ship:setWeaponTubeDirection(7, 120)				
-	ship:setWeaponTubeExclusiveFor(0,"Homing")
-	ship:setWeaponTubeExclusiveFor(1,"Homing")
-	ship:setWeaponTubeExclusiveFor(2,"Homing")
-	ship:setWeaponTubeExclusiveFor(3,"Homing")
-	ship:setWeaponTubeExclusiveFor(4,"Homing")
-	ship:setWeaponTubeExclusiveFor(5,"Homing")
-	ship:setWeaponTubeExclusiveFor(6,"Homing")
-	ship:setWeaponTubeExclusiveFor(7,"Homing")
-	ship:setWeaponStorageMax("Homing",32)						--more (vs 5)
-	ship:setWeaponStorage("Homing", 32)		
-	ship:setWeaponStorageMax("HVLI",0)							--less (vs 10)
-	ship:setWeaponStorage("HVLI", 0)
-	ship:setRadarTrace("missile_cruiser.png")				--different radar trace
-	local ships_key = _("scienceDB","Ships")
-	local frigate_key = _("scienceDB","Frigate")
-	local predator_key = _("scienceDB","Predator")
-	local pirahna_key = _("scienceDB","Piranha F8")
-	local predator_db = queryScienceDatabase(ships_key,frigate_key,predator_key)
-	if predator_db == nil then
-		local frigate_db = queryScienceDatabase(ships_key,frigate_key)
-		if frigate_db ~= nil then	--added for translation issues
-			frigate_db:addEntry(predator_key)
-			predator_db = queryScienceDatabase(ships_key,frigate_key,predator_key)
-			local tube_key = _("scienceDB","Large tube -60")
-			local tube2_key = _("scienceDB","Tube -90")
-			local tube3_key = _("scienceDB","Large tube -90")
-			local tube4_key = _("scienceDB","Large tube 60")
-			local tube5_key = _("scienceDB","Tube 90")
-			local tube6_key = _("scienceDB","Large tube 90")
-			local tube7_key = _("scienceDB","Tube -120")
-			local tube8_key = _("scienceDB","Tube 120")
-			local load_val = _("scienceDB","12 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,frigate_key,pirahna_key),	--base ship database entry
-				predator_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Predator is a significantly improved Piranha F8. Stronger shields and hull, faster impulse and turning speeds, a jump drive, beam weapons, eight missile tubes pointing in six directions and a large number of homing missiles to shoot."),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},		--torpedo tube direction and load speed
-					{key = tube3_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube4_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube5_key, value = load_val},		--torpedo tube direction and load speed
-					{key = tube6_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube7_key, value = load_val},		--torpedo tube direction and load speed
-					{key = tube8_key, value = load_val},		--torpedo tube direction and load speed
-				},
-				"5 - 35 U"		--jump range
-			)
-			predator_db:setImage("radar/missile_cruiser.png")		--override default radar image
-			predator_db:setModelDataName("HeavyCorvetteRed")
-		end
-	end
-	return ship		
-end
-function atlantisY42(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Atlantis X23")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Atlantis Y42")
-	ship:setShieldsMax(300,200,300,200)							--stronger shields (vs 200,200,200,200)
-	ship:setShields(300,200,300,200)					
-	ship:setImpulseMaxSpeed(65)									--faster impulse (vs 30)
-	ship:setRotationMaxSpeed(15)								--faster maneuver (vs 3.5)
---				   Index,  Arc,	  Dir, Range,	Cycle,	Damage
-	ship:setBeamWeapon(2,	80,	  190,	1500,		6,		 8)	--narrower (vs 100)
-	ship:setBeamWeapon(3,	80,	  170,	1500,		6,		 8)	--extra (vs 3 beams)
-	ship:setWeaponStorageMax("Homing",16)						--more (vs 4)
-	ship:setWeaponStorage("Homing", 16)
-	local ships_key = _("scienceDB","Ships")
-	local corvette_key = _("scienceDB","Corvette")
-	local y42_key = _("scienceDB","Atlantis Y42")
-	local atlantis_key = _("scienceDB","Atlantis X23")
-	local atlantis_y42_db = queryScienceDatabase(ships_key,corvette_key,y42_key)
-	if atlantis_y42_db == nil then
-		local corvette_db = queryScienceDatabase(ships_key,corvette_key)
-		if corvette_db ~= nil then	--added for translation issues
-			corvette_db:addEntry(y42_key)
-			atlantis_y42_db = queryScienceDatabase(ships_key,corvette_key,y42_key)
-			local tube_key = _("scienceDB","Tube -90")
-			local tube2_key = _("scienceDB"," Tube -90")
-			local tube3_key = _("scienceDB","Tube 90")
-			local tube4_key = _("scienceDB"," Tube 90")
-			local load_val = _("scienceDB","10 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,corvette_key,atlantis_key),	--base ship database entry
-				atlantis_y42_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Atlantis Y42 improves on the Atlantis X23 with stronger shields, faster impulse and turn speeds, an extra beam in back and a larger missile stock"),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube3_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube4_key, value = load_val},	--torpedo tube direction and load speed
-				},
-				"5 - 50 U"		--jump range
-			)
-		end
-	end
-	return ship		
-end
-function starhammerV(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Starhammer II")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Starhammer V")
-	ship:setImpulseMaxSpeed(65)									--faster impulse (vs 35)
-	ship:setRotationMaxSpeed(15)								--faster maneuver (vs 6)
-	ship:setShieldsMax(450, 350, 250, 250, 350)					--stronger shields (vs 450, 350, 150, 150, 350)
-	ship:setShields(450, 350, 250, 250, 350)					
---				   Index,  Arc,	  Dir, Range,	Cycle,	Damage
-	ship:setBeamWeapon(4,	60,	  180,	1500,		8,		11)	--extra rear facing beam
-	ship:setWeaponStorageMax("Homing",16)						--more (vs 4)
-	ship:setWeaponStorage("Homing", 16)		
-	ship:setWeaponStorageMax("HVLI",36)							--more (vs 20)
-	ship:setWeaponStorage("HVLI", 36)
-	local ships_key = _("scienceDB","Ships")
-	local corvette_key = _("scienceDB","Corvette")
-	local starhammerV_key = _("scienceDB","Starhammer V")
-	local starhammer2_key = _("scienceDB","Starhammer II")
-	local starhammer_v_db = queryScienceDatabase(ships_key,corvette_key,starhammerV_key)
-	if starhammer_v_db == nil then
-		local corvette_db = queryScienceDatabase(ships_key,corvette_key)
-		if corvette_db ~= nil then	--added for translation issues
-			corvette_db:addEntry(starhammerV_key)
-			starhammer_v_db = queryScienceDatabase(ships_key,corvette_key,starhammerV_key)
-			local tube_key = _("scienceDB","Tube 0")
-			local tube2_key = _("scienceDB"," Tube 0")
-			local load_val = _("scienceDB","10 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,corvette_key,starhammer2_key),	--base ship database entry
-				starhammer_v_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Starhammer V recognizes common modifications made in the field to the Starhammer II: stronger shields, faster impulse and turning speeds, additional rear beam and more missiles to shoot. These changes make the Starhammer V a force to be reckoned with."),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},	--torpedo tube direction and load speed
-				},
-				"5 - 50 U"		--jump range
-			)
-		end
-	end
-	return ship		
-end
-function tyr(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Battlestation")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Tyr")
-	ship:setImpulseMaxSpeed(50)									--faster impulse (vs 30)
-	ship:setRotationMaxSpeed(10)								--faster maneuver (vs 1.5)
-	ship:setShieldsMax(400, 300, 300, 400, 300, 300)			--stronger shields (vs 300, 300, 300, 300, 300)
-	ship:setShields(400, 300, 300, 400, 300, 300)					
-	ship:setHullMax(100)										--stronger hull (vs 70)
-	ship:setHull(100)
---				   Index,  Arc,	  Dir, Range,	Cycle,	Damage
-	ship:setBeamWeapon(0,	90,	  -60,	2500,		6,		 8)	--stronger beams, broader coverage
-	ship:setBeamWeapon(1,	90,	 -120,	2500,		6,		 8)
-	ship:setBeamWeapon(2,	90,	   60,	2500,		6,		 8)
-	ship:setBeamWeapon(3,	90,	  120,	2500,		6,		 8)
-	ship:setBeamWeapon(4,	90,	  -60,	2500,		6,		 8)
-	ship:setBeamWeapon(5,	90,	 -120,	2500,		6,		 8)
-	ship:setBeamWeapon(6,	90,	   60,	2500,		6,		 8)
-	ship:setBeamWeapon(7,	90,	  120,	2500,		6,		 8)
-	ship:setBeamWeapon(8,	90,	  -60,	2500,		6,		 8)
-	ship:setBeamWeapon(9,	90,	 -120,	2500,		6,		 8)
-	ship:setBeamWeapon(10,	90,	   60,	2500,		6,		 8)
-	ship:setBeamWeapon(11,	90,	  120,	2500,		6,		 8)
-	local ships_key = _("scienceDB","Ships")
-	local dreadnought_key = _("scienceDB","Dreadnought")
-	local tyr_key = _("scienceDB","Tyr")
-	local battlestation_key = _("scienceDB","Battlestation")
-	local tyr_db = queryScienceDatabase(ships_key,dreadnought_key,tyr_key)
-	if tyr_db == nil then
-		local corvette_db = queryScienceDatabase(ships_key,dreadnought_key)
-		if corvette_db ~= nil then	--added for translation issues
-			corvette_db:addEntry(tyr_key)
-			tyr_db = queryScienceDatabase(ships_key,dreadnought_key,tyr_key)
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,dreadnought_key,battlestation_key),	--base ship database entry
-				tyr_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Tyr is the shipyard's answer to admiral Konstatz' casual statement that the Battlestation model was too slow to be effective. The shipyards improved on the Battlestation by fitting the Tyr with more than twice the impulse speed and more than six times the turn speed. They threw in stronger shields and hull and wider beam coverage just to show that they could"),
-				nil,
-				"5 - 50 U"		--jump range
-			)
-		end
-	end
-	return ship
-end
-function gnat(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Ktlitan Drone")
-	ship:setTypeName("Gnat")
-	ship:setHullMax(15)					--weaker hull (vs 30)
-	ship:setHull(15)
-	ship:setImpulseMaxSpeed(140)		--faster impulse (vs 120)
-	ship:setRotationMaxSpeed(25)		--faster maneuver (vs 10)
---				   Index,  Arc,	  Dir, Range,	Cycle,	Damage
-	ship:setBeamWeapon(0,   40,		0,	 600,		4,		 3)	--weaker (vs 6) beam
-	local ships_key = _("scienceDB","Ships")
-	local no_class_key = _("scienceDB","No Class")
-	local gnat_key = _("scienceDB","Gnat")
-	local ktlitan_key = _("scienceDB","Ktlitan Drone")
-	local gnat_db = queryScienceDatabase(ships_key,no_class_key,gnat_key)
-	if gnat_db == nil then
-		local no_class_db = queryScienceDatabase(ships_key,no_class_key)
-		if no_class_db ~= nil then	--added for translation issues
-			no_class_db:addEntry(gnat_key)
-			gnat_db = queryScienceDatabase(ships_key,no_class_key,gnat_key)
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,no_class_key,ktlitan_key),	--base ship database entry
-				gnat_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Gnat is a nimbler version of the Ktlitan Drone. It's got half the hull, but it moves and turns faster"),
-				nil,
-				nil		--jump range
-			)
-		end
-	end
-	return ship
-end
-function cucaracha(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Tug")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Cucaracha")
-	ship:setShieldsMax(200, 50, 50, 50, 50, 50)		--stronger shields (vs 20)
-	ship:setShields(200, 50, 50, 50, 50, 50)					
-	ship:setHullMax(100)							--stronger hull (vs 50)
-	ship:setHull(100)
-	ship:setRotationMaxSpeed(20)					--faster maneuver (vs 10)
-	ship:setAcceleration(30)						--faster acceleration (vs 15)
---				   Index,  Arc,	  Dir, Range,	Cycle,	Damage
-	ship:setBeamWeapon(0,	60,	    0,	1500,		6,		10)	--extra rear facing beam
-	local ships_key = _("scienceDB","Ships")
-	local no_class_key = _("scienceDB","No Class")
-	local cucaracha_key = _("scienceDB","Cucaracha")
-	local tug_key = _("scienceDB","Tug")
-	local cucaracha_db = queryScienceDatabase(ships_key,no_class_key,cucaracha_key)
-	if cucaracha_db == nil then
-		local no_class_db = queryScienceDatabase(ships_key,no_class_key)
-		if no_class_db ~= nil then	--added for translation issues
-			no_class_db:addEntry(cucaracha_key)
-			cucaracha_db = queryScienceDatabase(ships_key,no_class_key,cucaracha_key)
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,no_class_key,tug_key),	--base ship database entry
-				cucaracha_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Cucaracha is a quick ship built around the Tug model with heavy shields and a heavy beam designed to be difficult to squash"),
-				nil,
-				nil		--jump range
-			)
-		end
-	end
-	return ship
-end
-function maniapak(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Adder MK5")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Maniapak")
-	ship:setRadarTrace("exuari_fighter.png")			--different radar trace
-	ship:setImpulseMaxSpeed(70)					--slower impulse (vs 80)
-	ship:setWeaponTubeCount(9)					--more (vs 1)
-	ship:setWeaponTubeDirection(0,  0)				
-	ship:setWeaponTubeDirection(1,-10)				
-	ship:setWeaponTubeDirection(2, 10)				
-	ship:setWeaponTubeDirection(3,  0)				
-	ship:setWeaponTubeDirection(4,-12)				
-	ship:setWeaponTubeDirection(5, 12)				
-	ship:setWeaponTubeDirection(6,  0)				
-	ship:setWeaponTubeDirection(7,-15)				
-	ship:setWeaponTubeDirection(8, 15)				
-	ship:setTubeSize(0,"small")
-	ship:setTubeSize(1,"small")
-	ship:setTubeSize(2,"small")
-	ship:setTubeSize(6,"large")
-	ship:setTubeSize(7,"large")
-	ship:setTubeSize(8,"large")
-	ship:setTubeLoadTime(0,15)
-	ship:setTubeLoadTime(1,16)
-	ship:setTubeLoadTime(2,17)
-	ship:setTubeLoadTime(3,18)
-	ship:setTubeLoadTime(4,19)
-	ship:setTubeLoadTime(5,20)
-	ship:setTubeLoadTime(6,21)
-	ship:setTubeLoadTime(7,22)
-	ship:setTubeLoadTime(8,23)
-	ship:setWeaponStorageMax("Homing", 27)		--more (vs 0)
-	ship:setWeaponStorage("Homing",    27)
-	ship:setWeaponStorageMax("EMP",    18)		--more (vs 0)
-	ship:setWeaponStorage("EMP",       18)
-	ship:setWeaponStorageMax("Nuke",   27)		--more (vs 0)
-	ship:setWeaponStorage("Nuke",      27)
-	ship:setWeaponStorageMax("HVLI",   36)		--more (vs 4)
-	ship:setWeaponStorage("HVLI",      36)
-	local ships_key = _("scienceDB","Ships")
-	local starfighter_key = _("scienceDB","Starfighter")
-	local maniapak_key = _("scienceDB","Maniapak")
-	local adder_key = _("scienceDB","Adder MK5")
-	local maniapak_db = queryScienceDatabase(ships_key,starfighter_key,maniapak_key)
-	if maniapak_db == nil then
-		local fighter_db = queryScienceDatabase(ships_key,starfighter_key)
-		if fighter_db ~= nil then
-			fighter_db:addEntry(maniapak_key)
-			maniapak_db = queryScienceDatabase(ships_key,starfighter_key,maniapak_key)
-			local tube_key = _("scienceDB","Small tube 0")
-			local tube2_key = _("scienceDB","Small tube -10")
-			local tube3_key = _("scienceDB","Small tube 10")
-			local tube4_key = _("scienceDB","Tube 0")
-			local tube5_key = _("scienceDB","Tube -12")
-			local tube6_key = _("scienceDB","Tube 12")
-			local tube7_key = _("scienceDB","Large tube 0")
-			local tube8_key = _("scienceDB","Large tube -15")
-			local tube9_key = _("scienceDB","Large tube 15")
-			local load_val = _("scienceDB","15 sec")
-			local load2_val = _("scienceDB","16 sec")
-			local load3_val = _("scienceDB","17 sec")
-			local load4_val = _("scienceDB","18 sec")
-			local load5_val = _("scienceDB","19 sec")
-			local load6_val = _("scienceDB","20 sec")
-			local load7_val = _("scienceDB","21 sec")
-			local load8_val = _("scienceDB","22 sec")
-			local load9_val = _("scienceDB","23 sec")
-			local storage_key = _("scienceDB","Missile Storage")
-			local storage_val = _("scienceDB","H:27 E:18 N:27 L:36")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,starfighter_key,adder_key),	--base ship database entry
-				maniapak_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Maniapak is an extreme modification of an Adder MK5 and a Blade. A maniacal designer was tasked with packing as many missiles as possible in this tiny starfighter frame. This record has yet to be beaten. Unfortunately, this ship is often a danger to friends as well as foes."),
-				{
-					{key = tube_key, value = load_val},		--torpedo tube size, direction and load speed
-					{key = tube2_key, value = load2_val},		--torpedo tube size, direction and load speed
-					{key = tube3_key, value = load3_val},		--torpedo tube size, direction and load speed
-					{key = tube4_key, value = load4_val},
-					{key = tube5_key, value = load5_val},
-					{key = tube6_key, value = load6_val},
-					{key = tube7_key, value = load7_val},
-					{key = tube8_key, value = load8_val},
-					{key = tube9_key, value = load9_val},
-					{key = storage_key, value = storage_val},
-				},
-				nil,
-				"AdlerLongRangeScoutYellow"
-			)
-			maniapak_db:setImage("radar/exuari_fighter.png")		--override default radar image
-		end
-	end
-	return ship		
-end
-function starhammerIII(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Starhammer II")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Starhammer III")
---				   Index,  Arc,	  Dir, Range,	Cycle,	Damage
-	ship:setBeamWeapon(4,	60,	  180,	1500,		8,		11)	--extra rear facing beam
-	ship:setTubeSize(0,"large")
-	ship:setWeaponStorageMax("Homing",16)						--more (vs 4)
-	ship:setWeaponStorage("Homing", 16)		
-	ship:setWeaponStorageMax("HVLI",36)							--more (vs 20)
-	ship:setWeaponStorage("HVLI", 36)
-	local ships_key = _("scienceDB","Ships")
-	local corvette_key = _("scienceDB","Corvette")
-	local starhammer3_key = _("scienceDB","Starhammer III")
-	local starhammer2_key = _("scienceDB","Starhammer II")
-	local starhammer_iii_db = queryScienceDatabase(ships_key,corvette_key,starhammer3_key)
-	if starhammer_iii_db == nil then
-		local corvette_db = queryScienceDatabase(ships_key,corvette_key)
-		if corvette_db ~= nil then	--added for translation issues
-			corvette_db:addEntry(starhammer3_key)
-			starhammer_iii_db = queryScienceDatabase(ships_key,corvette_key,starhammer3_key)
-			local tube_key = _("scienceDB","Large tube 0")
-			local tube2_key = _("scienceDB","Tube 0")
-			local load_val = _("scienceDB","10 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,corvette_key,starhammer2_key),	--base ship database entry
-				starhammer_iii_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The designers of the Starhammer III took the Starhammer II and added a rear facing beam, enlarged one of the missile tubes and added more missiles to fire"),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},			--torpedo tube direction and load speed
-				},
-				"5 - 50 U"		--jump range
-			)
-		end
-	end
-	return ship
-end
-function k2breaker(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Ktlitan Breaker")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("K2 Breaker")
-	ship:setHullMax(200)							--stronger hull (vs 120)
-	ship:setHull(200)
-	ship:setWeaponTubeCount(3)						--more (vs 1)
-	ship:setTubeSize(0,"large")						--large (vs normal)
-	ship:setWeaponTubeDirection(1,-30)				
-	ship:setWeaponTubeDirection(2, 30)
-	ship:setWeaponTubeExclusiveFor(0,"HVLI")		--only HVLI (vs any)
-	ship:setWeaponStorageMax("Homing",16)			--more (vs 0)
-	ship:setWeaponStorage("Homing", 16)
-	ship:setWeaponStorageMax("HVLI",8)				--more (vs 5)
-	ship:setWeaponStorage("HVLI", 8)
-	local ships_key = _("scienceDB","Ships")
-	local no_class_key = _("scienceDB","No Class")
-	local k2_key = _("scienceDB","K2 Breaker")
-	local ktlitan_key = _("scienceDB","Ktlitan Breaker")
-	local k2_breaker_db = queryScienceDatabase(ships_key,no_class_key,k2_key)
-	if k2_breaker_db == nil then
-		local no_class_db = queryScienceDatabase(ships_key,no_class_key)
-		if no_class_db ~= nil then	--added for translation issues
-			no_class_db:addEntry(k2_key)
-			k2_breaker_db = queryScienceDatabase(ships_key,no_class_key,k2_key)
-			local tube_key = _("scienceDB","Large tube 0")
-			local tube2_key = _("scienceDB","Tube -30")
-			local tube3_key = _("scienceDB","Tube 30")
-			local load_val = _("scienceDB","13 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,no_class_key,ktlitan_key),	--base ship database entry
-				k2_breaker_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The K2 Breaker designers took the Ktlitan Breaker and beefed up the hull, added two bracketing tubes, enlarged the center tube and added more missiles to shoot. Should be good for a couple of enemy ships"),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},		--torpedo tube direction and load speed
-					{key = tube3_key, value = load_val},		--torpedo tube direction and load speed
-				},
-				nil
-			)
-		end
-	end
-	return ship
-end
-function hurricane(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Piranha F8")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Hurricane")
-	ship:setJumpDrive(true)
-	ship:setJumpDriveRange(5000,40000)			
-	ship:setWeaponTubeCount(8)						--more (vs 3)
-	ship:setWeaponTubeExclusiveFor(1,"HVLI")		--only HVLI (vs any)
-	ship:setWeaponTubeDirection(1,  0)				--forward (vs -90)
-	ship:setTubeSize(3,"large")						
-	ship:setWeaponTubeDirection(3,-90)
-	ship:setTubeSize(4,"small")
-	ship:setWeaponTubeExclusiveFor(4,"Homing")
-	ship:setWeaponTubeDirection(4,-15)
-	ship:setTubeSize(5,"small")
-	ship:setWeaponTubeExclusiveFor(5,"Homing")
-	ship:setWeaponTubeDirection(5, 15)
-	ship:setWeaponTubeExclusiveFor(6,"Homing")
-	ship:setWeaponTubeDirection(6,-30)
-	ship:setWeaponTubeExclusiveFor(7,"Homing")
-	ship:setWeaponTubeDirection(7, 30)
-	ship:setWeaponStorageMax("Homing",24)			--more (vs 5)
-	ship:setWeaponStorage("Homing", 24)
-	local ships_key = _("scienceDB","Ships")
-	local frigate_key = _("scienceDB","Frigate")
-	local hurricane_key = _("scienceDB","Hurricane")
-	local pirahna_key = _("scienceDB","Piranha F8")
-	local hurricane_db = queryScienceDatabase(ships_key,frigate_key,hurricane_key)
-	if hurricane_db == nil then
-		local frigate_db = queryScienceDatabase(ships_key,frigate_key)
-		if frigate_db ~= nil then	--added for translation issues
-			frigate_db:addEntry(hurricane_key)
-			hurricane_db = queryScienceDatabase(ships_key,frigate_key,hurricane_key)
-			local tube_key = _("scienceDB","Large tube 0")
-			local tube2_key = _("scienceDB","Tube 0")
-			local tube3_key = _("scienceDB","Large tube 90")
-			local tube4_key = _("scienceDB","Large tube -90")
-			local tube5_key = _("scienceDB","Small tube -15")
-			local tube6_key = _("scienceDB","Small tube 15")
-			local tube7_key = _("scienceDB","Tube -30")
-			local tube8_key = _("scienceDB","Tube 30")
-			local load_val = _("scienceDB","12 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,frigate_key,pirahna_key),	--base ship database entry
-				hurricane_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Hurricane is designed to jump in and shower the target with missiles. It is based on the Piranha F8, but with a jump drive, five more tubes in various directions and sizes and lots more missiles to shoot"),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},			--torpedo tube direction and load speed
-					{key = tube3_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube4_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube5_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube6_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube7_key, value = load_val},		--torpedo tube direction and load speed
-					{key = tube8_key, value = load_val},		--torpedo tube direction and load speed
-				},
-				"5 - 40 U"		--jump range
-			)
-		end
-	end
-	return ship
-end
-function phobosT4(enemyFaction)
-	local ship = CpuShip():setFaction(enemyFaction):setTemplate("Phobos T3")
-	ship:onTakingDamage(function(self,instigator)
-		string.format("")	--serious proton needs a global context
-		if instigator ~= nil then
-			self.damage_instigator = instigator
-		end
-	end)
-	ship:setTypeName("Phobos T4")
-	ship:setRotationMaxSpeed(20)								--faster maneuver (vs 10)
-	ship:setShieldsMax(80,30)									--stronger shields (vs 50,40)
-	ship:setShields(80,30)					
---				   Index,  Arc,	  Dir, Range,	Cycle,	Damage
-	ship:setBeamWeapon(0,	90,	  -15,	1500,		6,		6)	--longer (vs 1200), faster (vs 8)
-	ship:setBeamWeapon(1,	90,	   15,	1500,		6,		6)	
-	local ships_key = _("scienceDB","Ships")
-	local frigate_key = _("scienceDB","Frigate")
-	local t4_key = _("scienceDB","Phobos T4")
-	local phobos_key = _("scienceDB","Phobos T3")
-	local phobos_t4_db = queryScienceDatabase(ships_key,frigate_key,t4_key)
-	if phobos_t4_db == nil then
-		local frigate_db = queryScienceDatabase(ships_key,frigate_key)
-		if frigate_db ~= nil then	--added for translation issues
-			frigate_db:addEntry(t4_key)
-			phobos_t4_db = queryScienceDatabase(ships_key,frigate_key,t4_key)
-			local tube_key = _("scienceDB","Tube -1")
-			local tube2_key = _("scienceDB","Tube 1")
-			local load_val = _("scienceDB","60 sec")
-			addShipToDatabase(
-				queryScienceDatabase(ships_key,frigate_key,phobos_key),	--base ship database entry
-				phobos_t4_db,	--modified ship database entry
-				ship,			--ship just created, long description on the next line
-				_("scienceDB","The Phobos T4 makes some simple improvements on the Phobos T3: faster maneuver, stronger front shields, though weaker rear shields and longer and faster beam weapons"),
-				{
-					{key = tube_key, value = load_val},	--torpedo tube direction and load speed
-					{key = tube2_key, value = load_val},		--torpedo tube direction and load speed
-				},
-				nil		--jump range
-			)
-		end
-	end
-	return ship
-end
-function serviceJonque(enemyFaction)
-	local ship = CpuShip():setTemplate("Garbage Jump Freighter 4")
-	if enemyFaction ~= nil then
-		ship:setFaction(enemyFaction)
-	end
-	local jonque_key = _("scienceDB","Service Jonque")
-	ship:setTypeName(jonque_key):setCommsScript(""):setCommsFunction(commsServiceJonque)
-	addFreighter(jonque_key,ship)	--update science database if applicable
-	return ship
-end
---	science database update functions
-function genericFreighterScienceInfo(specific_freighter_db,base_db,ship)
-	local freighter_key = _("scienceDB","Freighter")
-	specific_freighter_db:setImage("radar/transport.png")
-	specific_freighter_db:setKeyValue("Sub-class",freighter_key)
-	specific_freighter_db:setKeyValue("Size",base_db:getKeyValue("Size"))
-	local shields = ship:getShieldCount()
-	if shields > 0 then
-		local shield_string = ""
-		for i=1,shields do
-			if shield_string == "" then
-				shield_string = string.format("%i",math.floor(ship:getShieldMax(i-1)))
-			else
-				shield_string = string.format("%s/%i",shield_string,math.floor(ship:getShieldMax(i-1)))
-			end
-		end
-		specific_freighter_db:setKeyValue("Shield",shield_string)
-	end
-	specific_freighter_db:setKeyValue("Hull",string.format("%i",math.floor(ship:getHullMax())))
-	specific_freighter_db:setKeyValue("Move speed",string.format("%.1f u/min",ship:getImpulseMaxSpeed()*60/1000))
-	specific_freighter_db:setKeyValue("Turn speed",string.format("%.1f deg/sec",ship:getRotationMaxSpeed()))
-	if ship:hasJumpDrive() then
-		local base_jump_range = base_db:getKeyValue("Jump range")
-		if base_jump_range ~= nil and base_jump_range ~= "" then
-			specific_freighter_db:setKeyValue("Jump range",base_jump_range)
-		else
-			specific_freighter_db:setKeyValue("Jump range","5 - 50 u")
-		end
-	end
-	if ship:hasWarpDrive() then
-		specific_freighter_db:setKeyValue("Warp Speed",string.format("%.1f u/min",ship:getWarpSpeed()*60/1000))
-	end
-end
-function addFreighters()
-	local ships_key = _("scienceDB","Ships")
-	local freighter_key = _("scienceDB","Freighter")
-	local freighter_db = queryScienceDatabase(ships_key,freighter_key)
-	if freighter_db == nil then
-		local ship_db = queryScienceDatabase(ships_key)
-		ship_db:addEntry(freighter_key)
-		freighter_db = queryScienceDatabase(ships_key,freighter_key)
-		freighter_db:setImage("radar/transport.png")
-		freighter_db:setLongDescription(_("scienceDB","Small, medium and large scale transport ships. These are the working ships that keep commerce going in any sector. They may carry personnel, goods, cargo, equipment, garbage, fuel, research material, etc."))
-	end
-	return freighter_db
-end
-function addFreighter(freighter_type,ship)
-	local ships_key = _("scienceDB","Ships")
-	local freighter_key = _("scienceDB","Freighter")
-	local corvette_key = _("scienceDB","Corvette")
-	local sedan_key = _("scienceDB","Space Sedan")
-	local omnibus_key = _("scienceDB","Omnibus")
-	local jonque_key = _("scienceDB","Service Jonque")
-	local courier_key = _("scienceDB","Courier")
-	local wagon_key = _("scienceDB","Work Wagon")
-	local lorry_key = _("scienceDB","Laden Lorry")
-	local physics_key = _("scienceDB","Physics Research")
-	local freighter_db = addFreighters()
-	if freighter_type ~= nil then
-		if freighter_type == sedan_key then
-			local space_sedan_db = queryScienceDatabase(ships_key,freighter_key,sedan_key)
-			if space_sedan_db == nil then
-				local pjf3_key = _("scienceDB","Personnel Jump Freighter 3")
-				freighter_db:addEntry(sedan_key)
-				space_sedan_db = queryScienceDatabase(ships_key,freighter_key,sedan_key)
-				genericFreighterScienceInfo(space_sedan_db,queryScienceDatabase(ships_key,corvette_key,pjf3_key),ship)
-				space_sedan_db:setModelDataName("transport_1_3")
-				space_sedan_db:setLongDescription(_("scienceDB","The Space Sedan was built around a surplus Personnel Jump Freighter 3. It's designed to provide relatively low cost transportation primarily for people, but there is also a limited amount of cargo space available"))
-			end
-		elseif freighter_type == omnibus_key then
-			local omnibus_db = queryScienceDatabase(ships_key,freighter_key,omnibus_key)
-			if omnibus_db == nil then
-				local pjf5_key = _("scienceDB","Personnel Jump Freighter 5")
-				freighter_db:addEntry(omnibus_key)
-				omnibus_db = queryScienceDatabase(ships_key,freighter_key,omnibus_key)
-				genericFreighterScienceInfo(omnibus_db,queryScienceDatabase(ships_key,corvette_key,pjf5_key),ship)
-				omnibus_db:setModelDataName("transport_1_5")
-				omnibus_db:setLongDescription(_("scienceDB","The Omnibus was designed from the Personnel Jump Freighter 5. It's made to transport large numbers of passengers of various types along with their luggage and any associated cargo"))
-			end
-		elseif freighter_type == jonque_key then
-			local service_jonque_db = queryScienceDatabase(ships_key,freighter_key,jonque_key)
-			if service_jonque_db == nil then
-				local ejf4_key = _("scienceDB","Equipment Jump Freighter 4")
-				freighter_db:addEntry(jonque_key)
-				service_jonque_db = queryScienceDatabase(ships_key,freighter_key,jonque_key)
-				genericFreighterScienceInfo(service_jonque_db,queryScienceDatabase(ships_key,corvette_key,"Equipment Jump Freighter 4"),ship)
-				service_jonque_db:setModelDataName("transport_4_4")
-				service_jonque_db:setLongDescription(_("scienceDB","The Service Jonque is a modified Equipment Jump Freighter 4. It's designed to carry spare parts and equipment as well as the necessary repair personnel to where it's needed to repair stations and ships"))
-			end
-		elseif freighter_type == courier_key then
-			local courier_db = queryScienceDatabase(ships_key,freighter_key,courier_key)
-			if courier_db == nil then
-				local pf1_key = _("scienceDB","Personnel Freighter 1")
-				freighter_db:addEntry(courier_key)
-				courier_db = queryScienceDatabase(ships_key,freighter_key,courier_key)
-				genericFreighterScienceInfo(courier_db,queryScienceDatabase(ships_key,corvette_key,pf1_key),ship)
-				courier_db:setModelDataName("transport_1_1")
-				courier_db:setLongDescription(_("scienceDB","The Courier is a souped up Personnel Freighter 1. It's made to deliver people and messages fast. Very fast"))
-			end
-		elseif freighter_type == wagon_key then
-			local work_wagon_db = queryScienceDatabase(ships_key,freighter_key,wagon_key)
-			if work_wagon_db == nil then
-				local ef2_key = _("scienceDB","Equipment Freighter 2")
-				freighter_db:addEntry(wagon_key)
-				work_wagon_db = queryScienceDatabase(ships_key,freighter_key,wagon_key)
-				genericFreighterScienceInfo(work_wagon_db,queryScienceDatabase(ships_key,corvette_key,ef2_key),ship)
-				work_wagon_db:setModelDataName("transport_4_2")
-				work_wagon_db:setLongDescription(_("scienceDB","The Work Wagon is a conversion of an Equipment Freighter 2 designed to carry equipment and parts where they are needed for repair or construction."))
-			end
-		elseif freighter_type == lorry_key then
-			local laden_lorry_db = queryScienceDatabase(ships_key,freighter_key,lorry_key)
-			if laden_lorry_db == nil then
-				local gf3_key = _("scienceDB","Goods Freighter 3")
-				freighter_db:addEntry(lorry_key)
-				laden_lorry_db = queryScienceDatabase(ships_key,freighter_key,lorry_key)
-				genericFreighterScienceInfo(laden_lorry_db,queryScienceDatabase(ships_key,corvette_key,gf3_key),ship)
-				laden_lorry_db:setModelDataName("transport_2_3")
-				laden_lorry_db:setLongDescription(_("scienceDB","As a side contract, Conversion R Us put together the Laden Lorry from some recently acquired Goods Freighter 3 hulls. The added warp drive makes for a more versatile goods carrying vessel."))
-			end
-		elseif freighter_type == physics_key then
-			local physics_research_db = queryScienceDatabase(ships_key,freighter_key,physics_key)
-			if physics_research_db == nil then
-				local garf3_key = _("scienceDB","Garbage Freighter 3")
-				freighter_db:addEntry(physics_key)
-				physics_research_db = queryScienceDatabase(ships_key,freighter_key,physics_key)
-				genericFreighterScienceInfo(physics_research_db,queryScienceDatabase(ships_key,corvette_key,garf3_key),ship)
-				physics_research_db:setModelDataName("transport_3_3")
-				physics_research_db:setLongDescription(_("scienceDB","Conversion R Us cleaned up and converted excess freighter hulls into Physics Research vessels. The reduced weight improved the impulse speed and maneuverability."))
-			end
-		end
-	end
-end
-function addShipToDatabase(base_db,modified_db,ship,description,tube_directions,jump_range,model_name)
-	modified_db:setLongDescription(description)
-	if base_db ~= nil then
-		modified_db:setImage(base_db:getImage())
-		local class_key = _("scienceDB","Class")
-		local subclass_key = _("scienceDB","Sub-class")
-		local size_key = _("scienceDB","Size")
-		modified_db:setKeyValue(class_key,base_db:getKeyValue(class_key))
-		modified_db:setKeyValue(subclass_key,base_db:getKeyValue(subclass_key))
-		modified_db:setKeyValue(size_key,base_db:getKeyValue(size_key))
-	end
-	local shields = ship:getShieldCount()
-	if shields > 0 then
-		local shield_string = ""
-		for i=1,shields do
-			if shield_string == "" then
-				shield_string = string.format("%i",math.floor(ship:getShieldMax(i-1)))
-			else
-				shield_string = string.format("%s/%i",shield_string,math.floor(ship:getShieldMax(i-1)))
-			end
-		end
-		local shield_key = _("scienceDB","Shield")
-		modified_db:setKeyValue(shield_key,shield_string)
-	end
-	local hull_key = _("scienceDB","Hull")
-	local move_speed_key = _("scienceDB","Move speed")
-	local reverse_move_speed_key = _("scienceDB","Reverse move speed")
-	local turn_speed_key = _("scienceDB","Turn speed")
-	local impulse_forward, impulse_reverse = ship:getImpulseMaxSpeed()
-	modified_db:setKeyValue(hull_key,string.format("%i",math.floor(ship:getHullMax())))
-	modified_db:setKeyValue(move_speed_key,string.format(_("scienceDB","%.1f u/min"),impulse_forward*60/1000))
-	modified_db:setKeyValue(reverse_move_speed_key,string.format(_("scienceDB","%.1f u/min"),impulse_reverse*60/1000))
-	modified_db:setKeyValue(turn_speed_key,string.format(_("scienceDB","%.1f deg/sec"),ship:getRotationMaxSpeed()))
-	if ship:hasJumpDrive() then
-		local jump_range_key = _("scienceDB","Jump range")
-		if jump_range == nil then
-			local base_jump_range = nil
-			if base_db ~= nil then
-				base_jump_range = base_db:getKeyValue(jump_range_key)
-			end
-			if base_jump_range ~= nil and base_jump_range ~= "" then
-				modified_db:setKeyValue(jump_range_key,base_jump_range)
-			else
-				modified_db:setKeyValue(jump_range_key,"5 - 50 u")
-			end
-		else
-			modified_db:setKeyValue(jump_range_key,jump_range)
-		end
-	end
-	if ship:hasWarpDrive() then
-		local ward_speed_key = _("scienceDB","Warp Speed")
-		modified_db:setKeyValue(ward_speed_key,string.format(_("scienceDB","%.1f u/min"),ship:getWarpSpeed()*60/1000))
-	end
-	local key = ""
-	if ship:getBeamWeaponRange(0) > 0 then
-		local bi = 0
-		local count_repeat_loop = 0
-		repeat
-			local beam_direction = ship:getBeamWeaponDirection(bi)
-			if beam_direction > 315 and beam_direction < 360 then
-				beam_direction = beam_direction - 360
-			end
-			key = string.format(_("scienceDB","Beam weapon %.1f:%.1f"),ship:getBeamWeaponDirection(bi),ship:getBeamWeaponArc(bi))
-			while(modified_db:getKeyValue(key) ~= "") do
-				key = " " .. key
-			end
-			modified_db:setKeyValue(key,string.format(_("scienceDB","%.1f Dmg / %.1f sec"),ship:getBeamWeaponDamage(bi),ship:getBeamWeaponCycleTime(bi)))
-			bi = bi + 1
-			count_repeat_loop = count_repeat_loop + 1
-		until(ship:getBeamWeaponRange(bi) < 1 or count_repeat_loop > max_repeat_loop)
-		if count_repeat_loop > max_repeat_loop then
-			print("repeated too many times when going through beams")
-		end
-	end
-	local tubes = ship:getWeaponTubeCount()
-	if tubes > 0 then
-		if tube_directions ~= nil then
-			for i=1,#tube_directions do
-				modified_db:setKeyValue(tube_directions[i].key,tube_directions[i].value)
-			end
-		end
-		local missile_types = {'Homing', 'Nuke', 'Mine', 'EMP', 'HVLI'}
-		for index, missile_type in ipairs(missile_types) do
-			local max_storage = ship:getWeaponStorageMax(missile_type)
-			if max_storage > 0 then
-				modified_db:setKeyValue(string.format(_("scienceDB","Storage %s"),missile_type),string.format("%i",max_storage))
-			end
-		end
-	end
-	if model_name ~= nil then
-		modified_db:setModelDataName(model_name)
-	end
-end
 ------------------------
 --	Update functions  --
 ------------------------
